Data.Aeson.Extras

Description

Encoding and decoding of ByteString and serialisable values as base16 encoded JSON strings

Documentation

encodeByteString :: ByteString -> Text

decodeByteString :: Value -> Parser ByteString

encodeSerialise :: Serialise a => a -> Text

decodeSerialise :: Serialise a => Value -> Parser a

tryDecode :: Text -> Either String ByteString

newtype JSONViaSerialise a

Newtype for deriving ToJSON and FromJSON for types that have a Serialise instance by just encoding the serialized bytes as a JSON string.

Constructors

JSONViaSerialise a