Ledger

Documentation

data SatInt

newtype ExMemory

Constructors

ExMemory CostingInteger

data ExBudget

Constructors

ExBudget

Fields

exBudgetCPU :: ExCPU

exBudgetMemory :: ExMemory

newtype ExCPU

Constructors

ExCPU CostingInteger

findContinuingOutputs :: ScriptContext -> [Integer]

findDatum :: DatumHash -> TxInfo -> Maybe Datum

findDatumHash :: Datum -> TxInfo -> Maybe DatumHash

findOwnInput :: ScriptContext -> Maybe TxInInfo

findTxInByTxOutRef :: TxOutRef -> TxInfo -> Maybe TxInInfo

fromSymbol :: CurrencySymbol -> ValidatorHash

getContinuingOutputs :: ScriptContext -> [TxOut]

ownCurrencySymbol :: ScriptContext -> CurrencySymbol

ownHash :: ScriptContext -> ValidatorHash

ownHashes :: ScriptContext -> (ValidatorHash, DatumHash)

pubKeyOutput :: TxOut -> Maybe PubKeyHash

pubKeyOutputsAt :: PubKeyHash -> TxInfo -> [Value]

scriptOutputsAt :: ValidatorHash -> TxInfo -> [(DatumHash, Value)]

spendsOutput :: TxInfo -> TxId -> Integer -> Bool

txSignedBy :: TxInfo -> PubKeyHash -> Bool

valueLockedBy :: TxInfo -> ValidatorHash -> Value

valuePaidTo :: TxInfo -> PubKeyHash -> Value

valueProduced :: TxInfo -> Value

valueSpent :: TxInfo -> Value

data ScriptContext

Constructors

ScriptContext

Fields

scriptContextTxInfo :: TxInfo

scriptContextPurpose :: ScriptPurpose

data ScriptPurpose

Constructors

Minting CurrencySymbol

Spending TxOutRef

Rewarding StakingCredential

Certifying DCert

data TxInInfo

Constructors

TxInInfo

Fields

txInInfoOutRef :: TxOutRef

txInInfoResolved :: TxOut

data TxInfo

Constructors

TxInfo

Fields

txInfoInputs :: [TxInInfo]

txInfoOutputs :: [TxOut]

txInfoFee :: Value

txInfoMint :: Value

txInfoDCert :: [DCert]

txInfoWdrl :: [(StakingCredential, Integer)]

txInfoValidRange :: POSIXTimeRange

txInfoSignatories :: [PubKeyHash]

txInfoData :: [(DatumHash, Datum)]

txInfoId :: TxId

data TxOutRef

Constructors

TxOutRef

Fields

txOutRefId :: TxId

txOutRefIdx :: Integer

newtype POSIXTime

Constructors

POSIXTime

Fields

getPOSIXTime :: Integer

newtype PubKeyHash

Constructors

PubKeyHash

Fields

getPubKeyHash :: BuiltinByteString

newtype DatumHash

Constructors

DatumHash BuiltinByteString

newtype Datum

Constructors

Datum

Fields

getDatum :: BuiltinData

type POSIXTimeRange = Interval POSIXTime

data Interval a

Constructors

Interval

Fields

ivFrom :: LowerBound a

ivTo :: UpperBound a

newtype ValidatorHash

Constructors

ValidatorHash BuiltinByteString

toPubKeyHash :: Address -> Maybe PubKeyHash

newtype MintingPolicyHash

Constructors

MintingPolicyHash BuiltinByteString

inScripts :: TxIn -> Maybe (Validator, Redeemer, Datum)

isPayToScriptOut :: TxOut -> Bool

isPubKeyOut :: TxOut -> Bool

pubKeyHashTxOut :: Value -> PubKeyHash -> TxOut

pubKeyTxIns :: Fold (Set TxIn) TxIn

scriptTxIns :: Fold (Set TxIn) TxIn

data RedeemerPtr

Constructors

RedeemerPtr ScriptTag Integer

type Redeemers = Map RedeemerPtr Redeemer

data ScriptTag

Constructors

Spend

Mint

Cert

Reward

newtype Validator

Constructors

Validator

Fields

getValidator :: Script

newtype Redeemer

Constructors

Redeemer

Fields

getRedeemer :: BuiltinData

data Address

Constructors

Address

Fields

addressCredential :: Credential

addressStakingCredential :: Maybe StakingCredential

applyArguments :: Script -> [Data] -> Script

applyMintingPolicyScript :: Context -> MintingPolicy -> Redeemer -> Script

applyStakeValidatorScript :: Context -> StakeValidator -> Redeemer -> Script

applyValidator :: Context -> Validator -> Datum -> Redeemer -> Script

evaluateScript :: MonadError ScriptError m => Script -> m (ExBudget, [Text])

fromCompiledCode :: CompiledCode a -> Script

mkMintingPolicyScript :: CompiledCode (BuiltinData -> BuiltinData -> ()) -> MintingPolicy

mkStakeValidatorScript :: CompiledCode (BuiltinData -> BuiltinData -> ()) -> StakeValidator

mkValidatorScript :: CompiledCode (BuiltinData -> BuiltinData -> BuiltinData -> ()) -> Validator

runMintingPolicyScript :: MonadError ScriptError m => Context -> MintingPolicy -> Redeemer -> m (ExBudget, [Text])

runScript :: MonadError ScriptError m => Context -> Validator -> Datum -> Redeemer -> m (ExBudget, [Text])

runStakeValidatorScript :: MonadError ScriptError m => Context -> StakeValidator -> Redeemer -> m (ExBudget, [Text])

scriptSize :: Script -> Integer

unMintingPolicyScript :: MintingPolicy -> Script

unStakeValidatorScript :: StakeValidator -> Script

unValidatorScript :: Validator -> Script

unitDatum :: Datum

unitRedeemer :: Redeemer

newtype Context

Constructors

Context BuiltinData

newtype MintingPolicy

Constructors

MintingPolicy

Fields

getMintingPolicy :: Script

newtype RedeemerHash

Constructors

RedeemerHash BuiltinByteString

newtype Script

Constructors

Script

Fields

unScript :: Program DeBruijn DefaultUni DefaultFun ()

data ScriptError

Constructors

EvaluationError [Text] String

EvaluationException String String

newtype ScriptHash

Constructors

ScriptHash

Fields

getScriptHash :: BuiltinByteString

newtype StakeValidator

Constructors

StakeValidator

Fields

getStakeValidator :: Script

newtype StakeValidatorHash

Constructors

StakeValidatorHash BuiltinByteString

examplePlutusScriptAlwaysFails :: WitCtx witctx -> PlutusScript PlutusScriptV1

examplePlutusScriptAlwaysSucceeds :: WitCtx witctx -> PlutusScript PlutusScriptV1

filterValue :: (AssetId -> Bool) -> Value -> Value

negateValue :: Value -> Value

selectAsset :: Value -> AssetId -> Quantity

selectLovelace :: Value -> Lovelace

valueFromList :: [(AssetId, Quantity)] -> Value

valueToList :: Value -> [(AssetId, Quantity)]

data WitCtx witctx where

Constructors

WitCtxTxIn :: WitCtx WitCtxTxIn

WitCtxMint :: WitCtx WitCtxMint

WitCtxStake :: WitCtx WitCtxStake

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

data TxIn

Constructors

TxIn TxId TxIx

newtype TxIx

Constructors

TxIx Word

data AssetId

Constructors

AdaAssetId

AssetId !PolicyId !AssetName

data AssetName

newtype Lovelace

Constructors

Lovelace Integer

data PolicyId

data Value

dataHash :: BuiltinData -> BuiltinByteString

datumHash :: Datum -> DatumHash

mintingPolicyHash :: Versioned MintingPolicy -> MintingPolicyHash

redeemerHash :: Redeemer -> RedeemerHash

scriptCurrencySymbol :: Versioned MintingPolicy -> CurrencySymbol

scriptHash :: Versioned Script -> ScriptHash

stakeValidatorHash :: Versioned StakeValidator -> StakeValidatorHash

validatorHash :: Versioned Validator -> ValidatorHash

data Language

Constructors

PlutusV1

PlutusV2

data Versioned script

Constructors

Versioned

Fields

unversioned :: script

version :: Language

to :: a -> Interval a

from :: a -> Interval a

always :: Interval a

scriptHashAddress :: ValidatorHash -> Address

stakingCredential :: Address -> Maybe StakingCredential

toValidatorHash :: Address -> Maybe ValidatorHash

after :: Ord a => a -> Interval a -> Bool

before :: Ord a => a -> Interval a -> Bool

contains :: Ord a => Interval a -> Interval a -> Bool

hull :: Ord a => Interval a -> Interval a -> Interval a

intersection :: Ord a => Interval a -> Interval a -> Interval a

interval :: a -> a -> Interval a

isEmpty :: (Enum a, Ord a) => Interval a -> Bool

lowerBound :: a -> LowerBound a

member :: Ord a => a -> Interval a -> Bool

never :: Interval a

overlaps :: (Enum a, Ord a) => Interval a -> Interval a -> Bool

singleton :: a -> Interval a

strictLowerBound :: a -> LowerBound a

strictUpperBound :: a -> UpperBound a

upperBound :: a -> UpperBound a

type Closure = Bool

data Extended a

Constructors

NegInf

Finite a

PosInf

data LowerBound a

Constructors

LowerBound (Extended a) Closure

data UpperBound a

Constructors

UpperBound (Extended a) Closure

fromMilliSeconds :: DiffMilliSeconds -> POSIXTime

newtype DiffMilliSeconds

Constructors

DiffMilliSeconds Integer

examplePlutusScriptAlwaysSucceedsHash :: WitCtx ctx -> BuiltinByteString

examplePlutusScriptAlwaysFailsHash :: WitCtx ctx -> BuiltinByteString

newtype Slot

The slot number. This is a good proxy for time, since on the Cardano blockchain slots pass at a constant rate.

Constructors

Slot

Fields

getSlot :: Integer

type SlotRange = Interval Slot

An Interval of Slots.

width :: SlotRange -> Maybe Integer

Number of Slots covered by the interval, if finite. width (from x) == Nothing.

newtype Signature

A message with a cryptographic signature.

Constructors

Signature

Fields

getSignature :: BuiltinByteString

newtype Passphrase

Passphrase newtype to mark intent

Constructors

Passphrase

Fields

unPassphrase :: ByteString

newtype PubKey

Constructors

PubKey

Fields

getPubKey :: LedgerBytes

newtype PrivateKey

A cryptographic private key.

Constructors

PrivateKey

Fields

getPrivateKey :: LedgerBytes

pubKeyHash :: PubKey -> PubKeyHash

Compute the hash of a public key.

signedBy :: ByteArrayAccess a => Signature -> PubKey -> a -> Bool

Check whether the given Signature was signed by the private key corresponding to the given public key.

signTx :: TxId -> XPrv -> Passphrase -> Signature

Sign the hash of a transaction using a private key and passphrase.

signTx' :: TxId -> XPrv -> Signature

Sign the hash of a transaction using a private key that has no passphrase.

sign :: ByteArrayAccess a => a -> XPrv -> Passphrase -> Signature

Sign a message using a private key and passphrase.

sign' :: ByteArrayAccess a => a -> XPrv -> Signature

Sign a message using a private key with no passphrase.

generateFromSeed :: ByteString -> Passphrase -> XPrv

Generate a private key from a seed phrase and passphrase

generateFromSeed' :: ByteString -> XPrv

Generate a private key from a seed phrase without a passphrase.

xPubToPublicKey :: XPub -> PubKey

toPublicKey :: XPrv -> PubKey

newtype PaymentPubKey

Constructors

PaymentPubKey

Fields

unPaymentPubKey :: PubKey

newtype PaymentPrivateKey

Constructors

PaymentPrivateKey

Fields

unPaymentPrivateKey :: XPrv

type CardanoAddress = AddressInEra BabbageEra

cardanoAddressCredential :: AddressInEra era -> Credential

cardanoStakingCredential :: AddressInEra era -> Maybe StakingCredential

cardanoPubKeyHash :: AddressInEra era -> Maybe PubKeyHash

toPlutusAddress :: AddressInEra era -> Address

toPlutusPubKeyHash :: Hash PaymentKey -> PubKeyHash

newtype PaymentPubKeyHash

Constructors

PaymentPubKeyHash

Fields

unPaymentPubKeyHash :: PubKeyHash

xprvToPaymentPubKey :: XPrv -> PaymentPubKey

newtype StakePubKey

Constructors

StakePubKey

Fields

unStakePubKey :: PubKey

xprvToPaymentPubKeyHash :: XPrv -> PaymentPubKeyHash

newtype StakePubKeyHash

Constructors

StakePubKeyHash

Fields

unStakePubKeyHash :: PubKeyHash

xprvToStakePubKey :: XPrv -> StakePubKey

xprvToStakePubKeyHash :: XPrv -> StakePubKeyHash

xprvToStakingCredential :: XPrv -> StakingCredential

paymentPubKeyHash :: PaymentPubKey -> PaymentPubKeyHash

pubKeyHashAddress :: PaymentPubKeyHash -> Maybe StakingCredential -> Address

The address that should be targeted by a transaction output locked by the given public payment key (with its staking credentials).

pubKeyAddress :: PaymentPubKey -> Maybe StakingCredential -> Address

The address that should be targeted by a transaction output locked by the given public key. (with its staking credentials).

scriptValidatorHashAddress :: ValidatorHash -> Maybe StakingCredential -> Address

The address that should be used by a transaction output locked by the given validator script (with its staking credentials).

stakePubKeyHashCredential :: StakePubKeyHash -> StakingCredential

Construct a StakingCredential from a public key hash.

stakeValidatorHashCredential :: StakeValidatorHash -> StakingCredential

Construct a StakingCredential from a validator script hash.

mkValidatorCardanoAddress :: NetworkId -> Versioned Validator -> AddressInEra BabbageEra

Cardano address of a versioned Validator script.

data ToCardanoError

Constructors

TxBodyError String
A C.TxBodyError converted to String

DeserialisationError

InvalidValidityRange

ValueNotPureAda

OutputHasZeroAda

StakingPointersNotSupported

SimpleScriptsNotSupportedToCardano

MissingInputValidator

MissingDatum

MissingMintingPolicy

ScriptPurposeNotSupported ScriptTag

MissingMintingPolicyRedeemer

MissingStakeValidator

UnsupportedPlutusVersion Language

Tag String ToCardanoError

data CardanoTx where

Cardano tx from any era.

Constructors

CardanoTx :: IsCardanoEra era => Tx era -> EraInMode era CardanoMode -> CardanoTx

Bundled Patterns

pattern CardanoEmulatorEraTx :: Tx BabbageEra -> CardanoTx

adaToCardanoValue :: Ada -> Value

fromCardanoValue :: Value -> Value

toCardanoValue :: Value -> Either ToCardanoError Value

fromCardanoAssetId :: AssetId -> AssetClass

toCardanoAssetId :: AssetClass -> Either ToCardanoError AssetId

lovelaceToValue :: Lovelace -> Value

lovelaceValueOf :: Integer -> Value

adaValueOf :: Rational -> Value

isZero :: Value -> Bool

isAdaOnlyValue :: Value -> Bool

noAdaValue :: Value -> Value

adaOnlyValue :: Value -> Value

assetIdValue :: AssetId -> Integer -> Value

scale :: Integer -> Value -> Value

split :: Value -> (Value, Value)

policyId :: Versioned MintingPolicy -> PolicyId

combine :: Monoid m => (AssetId -> Quantity -> Quantity -> m) -> Value -> Value -> m

valueGeq :: Value -> Value -> Bool

valueLeq :: Value -> Value -> Bool

type ReferenceScript = ReferenceScript BabbageEra

type MintingWitnessesMap = Map MintingPolicyHash (Redeemer, Maybe (Versioned TxOutRef))

type ScriptsMap = Map ScriptHash (Versioned Script)

newtype TxOut

Constructors

TxOut

Fields

getTxOut :: TxOut CtxTx BabbageEra

data Certificate

Constructors

Certificate

Fields

certificateDcert :: DCert

certificateRedeemer :: Maybe Redeemer

redeemer for script credential

data Withdrawal

Stake withdrawal, if applicable the script should be included in txScripts.

Constructors

Withdrawal

Fields

withdrawalCredential :: Credential

staking credential

withdrawalAmount :: Integer

amount of withdrawal in Lovelace, must withdraw all eligible amount

withdrawalRedeemer :: Maybe Redeemer

redeemer for script credential

cardanoTxOutValue :: TxOut ctx era -> Value

txOutValue :: TxOut -> Value

outValue :: Lens TxOut TxOut Value (TxOutValue BabbageEra)

outValue' :: Lens' TxOut (TxOutValue BabbageEra)

toSizedTxOut :: TxOut -> Sized (TxOut StandardBabbage)

toCtxUTxOTxOut :: TxOut -> TxOut CtxUTxO BabbageEra

txOutDatumHash :: TxOut -> Maybe DatumHash

Get a hash from the stored TxOutDatum (either directly or by hashing the inlined datum)

txOutDatum :: forall d. FromData d => TxOut -> Maybe d

cardanoTxOutDatumHash :: TxOutDatum CtxUTxO BabbageEra -> Maybe (Hash ScriptData)

txOutPubKey :: TxOut -> Maybe PubKeyHash

txOutAddress :: TxOut -> CardanoAddress

outAddress :: Lens' TxOut (AddressInEra BabbageEra)

outDatumHash :: Lens TxOut TxOut (Maybe DatumHash) (TxOutDatum CtxTx BabbageEra)

txOutReferenceScript :: TxOut -> ReferenceScript

outReferenceScript :: Lens' TxOut ReferenceScript

lookupScript :: ScriptsMap -> ScriptHash -> Maybe (Versioned Script)

lookupValidator :: ScriptsMap -> ValidatorHash -> Maybe (Versioned Validator)

lookupMintingPolicy :: ScriptsMap -> MintingPolicyHash -> Maybe (Versioned MintingPolicy)

lookupStakeValidator :: ScriptsMap -> StakeValidatorHash -> Maybe (Versioned StakeValidator)

emptyTxBodyContent :: TxBodyContent BuildTx BabbageEra

data ValidationError

A reason why a transaction is invalid.

Constructors

TxOutRefNotFound TxIn
The transaction output consumed by a transaction input could not be found (either because it was already spent, or because there was no transaction with the given hash on the blockchain).

ScriptFailure ScriptError
For pay-to-script outputs: evaluation of the validator script failed.

CardanoLedgerValidationError Text
An error from Cardano.Ledger validation

MaxCollateralInputsExceeded
Balancing failed, it needed more than the maximum number of collateral inputs

type UtxoIndex = UTxO BabbageEra

The UTxOs of a blockchain indexed by their references.

newtype OnChainTx

A transaction on the blockchain. Invalid transactions are still put on the chain to be able to collect fees.

Constructors

OnChainTx

Fields

getOnChainTx :: Validated (Tx EmulatorEra)

eitherTx :: (CardanoTx -> r) -> (CardanoTx -> r) -> OnChainTx -> r

unOnChain :: OnChainTx -> CardanoTx

data ValidationResult

Constructors

FailPhase1 !CardanoTx !ValidationError
A transaction failed to validate in phase 1.

FailPhase2 !OnChainTx !ValidationError !Value
A transaction failed to validate in phase 2. The Value indicates the amount of collateral stored in the transaction.

Success !OnChainTx !RedeemerReport

type RedeemerReport = Map RdmrPtr ([Text], ExUnits)

type ValidationSuccess = (RedeemerReport, Validated (Tx EmulatorEra))

type ValidationErrorInPhase = (ValidationPhase, ValidationError)

data ValidationPhase

Constructors

Phase1

Phase2

_TxOutRefNotFound :: Prism' ValidationError TxIn

_ScriptFailure :: Prism' ValidationError ScriptError

_CardanoLedgerValidationError :: Prism' ValidationError Text

_FailPhase1 :: Prism' ValidationResult (CardanoTx, ValidationError)

_FailPhase2 :: Prism' ValidationResult (OnChainTx, ValidationError, Value)

_Success :: Prism' ValidationResult (OnChainTx, RedeemerReport)

cardanoTxFromValidationResult :: ValidationResult -> CardanoTx

toOnChain :: ValidationResult -> Maybe OnChainTx

getEvaluationLogs :: ValidationResult -> [Text]

Get logs from evaluating plutus scripts.

data DatumFromQuery

A datum in a transaction output that comes from a chain index query.

Constructors

DatumUnknown

DatumInline Datum

DatumInBody Datum

data DecoratedTxOut

Offchain view of a transaction output.

Constructors

PublicKeyDecoratedTxOut

Fields

_decoratedTxOutPubKeyHash :: PubKeyHash

The pubKey hash that protects the transaction address

_decoratedTxOutStakingCredential :: Maybe StakingCredential

The staking credential of the transaction address, if any

_decoratedTxOutValue :: Value

Value of the transaction output.

_decoratedTxOutPubKeyDatum :: Maybe (DatumHash, DatumFromQuery)

Optional datum (inline datum or datum in transaction body) attached to the transaction output.

_decoratedTxOutReferenceScript :: Maybe (Versioned Script)

Value of the transaction output.

ScriptDecoratedTxOut

Fields

_decoratedTxOutValidatorHash :: ValidatorHash

The hash of the script that protects the transaction address

_decoratedTxOutStakingCredential :: Maybe StakingCredential

The staking credential of the transaction address, if any

_decoratedTxOutValue :: Value

Value of the transaction output.

_decoratedTxOutScriptDatum :: (DatumHash, DatumFromQuery)

Datum attached to the transaction output, either in full (inline datum or datum in transaction body) or as a hash reference. A transaction output protected by a Plutus script is guardateed to have an associated datum.

_decoratedTxOutReferenceScript :: Maybe (Versioned Script)

Value of the transaction output.

_decoratedTxOutValidator :: Maybe (Versioned Validator)

Full version of the validator protecting the transaction output

datumInDatumFromQuery :: Traversal' DatumFromQuery Datum

decoratedTxOutPubKeyDatum :: Traversal' DecoratedTxOut (Maybe (DatumHash, DatumFromQuery))

decoratedTxOutPubKeyHash :: Traversal' DecoratedTxOut PubKeyHash

decoratedTxOutReferenceScript :: Lens' DecoratedTxOut (Maybe (Versioned Script))

decoratedTxOutScriptDatum :: Traversal' DecoratedTxOut (DatumHash, DatumFromQuery)

decoratedTxOutStakingCredential :: Lens' DecoratedTxOut (Maybe StakingCredential)

decoratedTxOutValidator :: Traversal' DecoratedTxOut (Maybe (Versioned Validator))

decoratedTxOutValidatorHash :: Traversal' DecoratedTxOut ValidatorHash

decoratedTxOutValue :: Lens' DecoratedTxOut Value

_PublicKeyDecoratedTxOut :: Prism' DecoratedTxOut (PubKeyHash, Maybe StakingCredential, Value, Maybe (DatumHash, DatumFromQuery), Maybe (Versioned Script))

_ScriptDecoratedTxOut :: Prism' DecoratedTxOut (ValidatorHash, Maybe StakingCredential, Value, (DatumHash, DatumFromQuery), Maybe (Versioned Script), Maybe (Versioned Validator))

mkDecoratedTxOut :: CardanoAddress -> Value -> Maybe (DatumHash, DatumFromQuery) -> Maybe (Versioned Script) -> Maybe DecoratedTxOut

mkPubkeyDecoratedTxOut :: CardanoAddress -> Value -> Maybe (DatumHash, DatumFromQuery) -> Maybe (Versioned Script) -> Maybe DecoratedTxOut

mkScriptDecoratedTxOut :: CardanoAddress -> Value -> (DatumHash, DatumFromQuery) -> Maybe (Versioned Script) -> Maybe (Versioned Validator) -> Maybe DecoratedTxOut

_decoratedTxOutAddress :: DecoratedTxOut -> Address

decoratedTxOutAddress :: Getter DecoratedTxOut Address

decoratedTxOutDatum :: Traversal' DecoratedTxOut (DatumHash, DatumFromQuery)

toDecoratedTxOut :: TxOut -> Maybe DecoratedTxOut

toTxOut :: NetworkId -> DecoratedTxOut -> Either ToCardanoError TxOut

toTxInfoTxOut :: DecoratedTxOut -> TxOut

Converts a transaction output from the chain index to the plutus-ledger-api transaction output.

Note that DecoratedTxOut supports features such inline datums and reference scripts which are not supported by V1 TxOut. Converting from DecoratedTxOut to TxOut and back is therefore lossy.

fromDecoratedIndex :: NetworkId -> Map TxOutRef DecoratedTxOut -> Either ToCardanoError UtxoIndex

getCardanoTxId :: CardanoTx -> TxId

getCardanoTxInputs :: CardanoTx -> [TxIn]

getTxBodyContentInputs :: TxBodyContent ctx era -> [TxIn]

getCardanoTxCollateralInputs :: CardanoTx -> [TxIn]

getTxBodyContentCollateralInputs :: TxBodyContent ctx era -> [TxIn]

getCardanoTxOutRefs :: CardanoTx -> [(TxOut, TxIn)]

getCardanoTxOutputs :: CardanoTx -> [TxOut]

getCardanoTxProducedOutputs :: CardanoTx -> Map TxIn TxOut

getCardanoTxSpentOutputs :: CardanoTx -> Set TxIn

getCardanoTxReturnCollateral :: CardanoTx -> Maybe TxOut

getTxBodyContentReturnCollateral :: TxBodyContent ctx BabbageEra -> Maybe TxOut

getCardanoTxProducedReturnCollateral :: CardanoTx -> Map TxIn TxOut

getCardanoTxTotalCollateral :: CardanoTx -> Maybe Lovelace

getCardanoTxFee :: CardanoTx -> Lovelace

getCardanoTxMint :: CardanoTx -> Value

getTxBodyContentMint :: TxBodyContent ctx era -> Value

getCardanoTxValidityRange :: CardanoTx -> SlotRange

getCardanoTxData :: CardanoTx -> Map DatumHash Datum

txBodyContentIns :: Lens' (TxBodyContent BuildTx BabbageEra) [(TxIn, BuildTxWith BuildTx (Witness WitCtxTxIn BabbageEra))]

txBodyContentCollateralIns :: Lens' (TxBodyContent BuildTx BabbageEra) [TxIn]

txBodyContentOuts :: Lens' (TxBodyContent ctx BabbageEra) [TxOut]

getCardanoTxRedeemers :: CardanoTx -> Redeemers

addCardanoTxSignature :: PrivateKey -> CardanoTx -> CardanoTx

decoratedTxOutPlutusValue :: DecoratedTxOut -> Value

type Blockchain = [Block]

A blockchain, which is just a list of blocks, starting with the newest.

type Block = [OnChainTx]

A block on the blockchain. This is just a list of transactions following on from the chain so far.

newtype BlockId

Block identifier (usually a hash)

Constructors

BlockId

Fields

getBlockId :: ByteString

onChainTxIsValid :: OnChainTx -> Bool

consumableInputs :: OnChainTx -> [TxIn]

Outputs consumed from the UTXO set by the OnChainTx

outputsProduced :: OnChainTx -> Map TxIn TxOut

Outputs added to the UTXO set by the OnChainTx

initialise :: Blockchain -> UtxoIndex

Create an index of all UTxOs on the chain.

insert :: CardanoTx -> UtxoIndex -> UtxoIndex

Update the index for the addition of a transaction.

insertCollateral :: CardanoTx -> UtxoIndex -> UtxoIndex

Update the index for the addition of only the collateral inputs of a failed transaction.

insertBlock :: Block -> UtxoIndex -> UtxoIndex

Update the index for the addition of a block.

lookup :: TxIn -> UtxoIndex -> Maybe TxOut

Find an unspent transaction output by the TxOutRef that spends it.

getCollateral :: UtxoIndex -> CardanoTx -> Value

adjustTxOut :: PParams (BabbageEra StandardCrypto) -> TxOut -> ([Lovelace], TxOut)

Adjust a single transaction output so it contains at least the minimum amount of Ada and return the adjustment (if any) and the updated TxOut.

minAdaTxOut :: PParams (BabbageEra StandardCrypto) -> TxOut -> Lovelace

Exact computation of the mimimum Ada required for a given TxOut. TODO: Should be moved to cardano-api-extended once created

minAdaTxOutEstimated :: Ada

Provide a reasonable estimate of the mimimum of Ada required for a TxOut.

An exact estimate of the the mimimum of Ada in a TxOut is determined by two things: - the PParams, more precisely its coinPerUTxOWord parameter. - the size of the TxOut. In many situations though, we need to determine a plausible value for the minimum of Ada needed for a TxOut without knowing much of the TxOut. This function provides a value big enough to balance UTxOs without a large inlined data (larger than a hash) nor a complex val with a lot of minted values. It's superior to the lowest minimum needed for an UTxO, as the lowest value require no datum. An estimate of the minimum required Ada for each tx output.

minLovelaceTxOutEstimated :: Lovelace

maxMinAdaTxOut :: Ada

maxFee :: Ada

TODO Should be calculated based on the maximum script size permitted on the Cardano blockchain.

genesisTxIn :: TxIn

cardano-ledger validation rules require the presence of inputs and we have to provide a stub TxIn for the genesis transaction.

createGenesisTransaction :: Map CardanoAddress Value -> CardanoTx

data DCert

data NetworkId

data Credential

data StakingCredential