Ledger.Address

Documentation

stakingCredential :: Address -> Maybe StakingCredential

The staking credential of an address (if any)

scriptHashAddress :: ValidatorHash -> Address

The address that should be used by a transaction output locked by the given validator script hash.

toValidatorHash :: Address -> Maybe ValidatorHash

The validator hash of the address, if any

toPubKeyHash :: Address -> Maybe PubKeyHash

The PubKeyHash of the address, if any

data Address

Address with two kinds of credentials, normal and staking.

Constructors

Address

Fields

addressCredential :: Credential

addressStakingCredential :: Maybe StakingCredential

type CardanoAddress = AddressInEra BabbageEra

newtype PaymentPrivateKey

Constructors

PaymentPrivateKey

Fields

unPaymentPrivateKey :: XPrv

newtype PaymentPubKey

Constructors

PaymentPubKey

Fields

unPaymentPubKey :: PubKey

newtype PaymentPubKeyHash

Constructors

PaymentPubKeyHash

Fields

unPaymentPubKeyHash :: PubKeyHash

newtype StakePubKey

Constructors

StakePubKey

Fields

unStakePubKey :: PubKey

newtype StakePubKeyHash

Constructors

StakePubKeyHash

Fields

unStakePubKeyHash :: PubKeyHash

toPlutusAddress :: AddressInEra era -> Address

toPlutusPubKeyHash :: Hash PaymentKey -> PubKeyHash

cardanoAddressCredential :: AddressInEra era -> Credential

cardanoPubKeyHash :: AddressInEra era -> Maybe PubKeyHash

cardanoStakingCredential :: AddressInEra era -> Maybe StakingCredential

paymentPubKeyHash :: PaymentPubKey -> PaymentPubKeyHash

pubKeyHashAddress :: PaymentPubKeyHash -> Maybe StakingCredential -> Address

The address that should be targeted by a transaction output locked by the given public payment key (with its staking credentials).

pubKeyAddress :: PaymentPubKey -> Maybe StakingCredential -> Address

The address that should be targeted by a transaction output locked by the given public key. (with its staking credentials).

scriptValidatorHashAddress :: ValidatorHash -> Maybe StakingCredential -> Address

The address that should be used by a transaction output locked by the given validator script (with its staking credentials).

stakePubKeyHashCredential :: StakePubKeyHash -> StakingCredential

Construct a StakingCredential from a public key hash.

stakeValidatorHashCredential :: StakeValidatorHash -> StakingCredential

Construct a StakingCredential from a validator script hash.

xprvToPaymentPubKey :: XPrv -> PaymentPubKey

xprvToPaymentPubKeyHash :: XPrv -> PaymentPubKeyHash

xprvToStakingCredential :: XPrv -> StakingCredential

xprvToStakePubKey :: XPrv -> StakePubKey

xprvToStakePubKeyHash :: XPrv -> StakePubKeyHash

mkValidatorCardanoAddress :: NetworkId -> Versioned Validator -> AddressInEra BabbageEra

Cardano address of a versioned Validator script.

Orphan instances

ToJSONKey (AddressInEra BabbageEra)

Instance details

Methods

toJSONKey :: ToJSONKeyFunction (AddressInEra BabbageEra)

toJSONKeyList :: ToJSONKeyFunction [AddressInEra BabbageEra]

FromJSONKey (AddressInEra BabbageEra)

Instance details

Methods

fromJSONKey :: FromJSONKeyFunction (AddressInEra BabbageEra)

fromJSONKeyList :: FromJSONKeyFunction [AddressInEra BabbageEra]