Ledger.AddressMap

Description

AddressMaps and functions for working on them.

AddressMaps are used to represent the limited knowledge about the state of the ledger that the wallet retains. Rather than keeping the entire ledger (which can be very large) the wallet only tracks the UTxOs at particular addresses.

Documentation

newtype AddressMap

A map of Addresses and their unspent outputs.

Constructors

AddressMap

Fields

getAddressMap :: Map CardanoAddress UtxoMap

type UtxoMap = Map TxIn (CardanoTx, TxOut)

addAddress :: CardanoAddress -> AddressMap -> AddressMap

Add an address with no unspent outputs to a map. If the address already exists, do nothing.

addAddresses :: [CardanoAddress] -> AddressMap -> AddressMap

Add a list of Addresses with no unspent outputs to the map.

fundsAt :: CardanoAddress -> Lens' AddressMap UtxoMap

Get the funds available at a particular address.

values :: AddressMap -> Map CardanoAddress Value

The total value of unspent outputs (which the map knows about) at an address.

traverseWithKey :: Applicative f => (CardanoAddress -> Map TxIn (CardanoTx, TxOut) -> f (Map TxIn (CardanoTx, TxOut))) -> AddressMap -> f AddressMap

Walk through the address map, applying an effectful function to each entry.

singleton :: (CardanoAddress, TxIn, CardanoTx, TxOut) -> AddressMap

An address map with a single unspent transaction output.

updateAddresses :: OnChainTx -> AddressMap -> AddressMap

Update an AddressMap with the inputs and outputs of a new transaction. updateAddresses does not add or remove any keys from the map.

updateAllAddresses :: OnChainTx -> AddressMap -> AddressMap

Update an AddressMap with the inputs and outputs of a new transaction, including all addresses in the transaction.

lookupOutRef :: TxIn -> AddressMap -> Maybe TxOut

Determine the unspent output that an input refers to

fromChain :: Blockchain -> AddressMap

The unspent transaction outputs of the ledger as a whole.