Ledger.Index

Description

An index of unspent transaction outputs, and some functions for validating transactions using the index.

Types for transaction validation based on UTXO index

type UtxoIndex = UTxO BabbageEra

The UTxOs of a blockchain indexed by their references.

insert :: CardanoTx -> UtxoIndex -> UtxoIndex

Update the index for the addition of a transaction.

insertCollateral :: CardanoTx -> UtxoIndex -> UtxoIndex

Update the index for the addition of only the collateral inputs of a failed transaction.

insertBlock :: Block -> UtxoIndex -> UtxoIndex

Update the index for the addition of a block.

initialise :: Blockchain -> UtxoIndex

Create an index of all UTxOs on the chain.

lookup :: TxIn -> UtxoIndex -> Maybe TxOut

Find an unspent transaction output by the TxOutRef that spends it.

getCollateral :: UtxoIndex -> CardanoTx -> Value

data ValidationError

A reason why a transaction is invalid.

Constructors

TxOutRefNotFound TxIn
The transaction output consumed by a transaction input could not be found (either because it was already spent, or because there was no transaction with the given hash on the blockchain).

ScriptFailure ScriptError
For pay-to-script outputs: evaluation of the validator script failed.

CardanoLedgerValidationError Text
An error from Cardano.Ledger validation

MaxCollateralInputsExceeded
Balancing failed, it needed more than the maximum number of collateral inputs

_TxOutRefNotFound :: Prism' ValidationError TxIn

_ScriptFailure :: Prism' ValidationError ScriptError

_CardanoLedgerValidationError :: Prism' ValidationError Text

data ValidationResult

Constructors

FailPhase1 !CardanoTx !ValidationError
A transaction failed to validate in phase 1.

FailPhase2 !OnChainTx !ValidationError !Value
A transaction failed to validate in phase 2. The Value indicates the amount of collateral stored in the transaction.

Success !OnChainTx !RedeemerReport

_Success :: Prism' ValidationResult (OnChainTx, RedeemerReport)

_FailPhase1 :: Prism' ValidationResult (CardanoTx, ValidationError)

_FailPhase2 :: Prism' ValidationResult (OnChainTx, ValidationError, Value)

cardanoTxFromValidationResult :: ValidationResult -> CardanoTx

toOnChain :: ValidationResult -> Maybe OnChainTx

getEvaluationLogs :: ValidationResult -> [Text]

Get logs from evaluating plutus scripts.

type ValidationSuccess = (RedeemerReport, Validated (Tx EmulatorEra))

type ValidationErrorInPhase = (ValidationPhase, ValidationError)

data ValidationPhase

Constructors

Phase1

Phase2

type RedeemerReport = Map RdmrPtr ([Text], ExUnits)

maxFee :: Ada

TODO Should be calculated based on the maximum script size permitted on the Cardano blockchain.

adjustTxOut :: PParams (BabbageEra StandardCrypto) -> TxOut -> ([Lovelace], TxOut)

Adjust a single transaction output so it contains at least the minimum amount of Ada and return the adjustment (if any) and the updated TxOut.

minAdaTxOut :: PParams (BabbageEra StandardCrypto) -> TxOut -> Lovelace

Exact computation of the mimimum Ada required for a given TxOut. TODO: Should be moved to cardano-api-extended once created

minAdaTxOutEstimated :: Ada

Provide a reasonable estimate of the mimimum of Ada required for a TxOut.

An exact estimate of the the mimimum of Ada in a TxOut is determined by two things: - the PParams, more precisely its coinPerUTxOWord parameter. - the size of the TxOut. In many situations though, we need to determine a plausible value for the minimum of Ada needed for a TxOut without knowing much of the TxOut. This function provides a value big enough to balance UTxOs without a large inlined data (larger than a hash) nor a complex val with a lot of minted values. It's superior to the lowest minimum needed for an UTxO, as the lowest value require no datum. An estimate of the minimum required Ada for each tx output.

minLovelaceTxOutEstimated :: Lovelace

maxMinAdaTxOut :: Ada

createGenesisTransaction :: Map CardanoAddress Value -> CardanoTx

genesisTxIn :: TxIn

cardano-ledger validation rules require the presence of inputs and we have to provide a stub TxIn for the genesis transaction.

data ExBudget

Constructors

ExBudget

Fields

exBudgetCPU :: ExCPU

exBudgetMemory :: ExMemory

newtype ExCPU

Constructors

ExCPU CostingInteger

newtype ExMemory

Constructors

ExMemory CostingInteger

data SatInt