Ledger.Index.Internal

Documentation

type UtxoIndex = UTxO BabbageEra

The UTxOs of a blockchain indexed by their references.

data ValidationError

A reason why a transaction is invalid.

Constructors

TxOutRefNotFound TxIn
The transaction output consumed by a transaction input could not be found (either because it was already spent, or because there was no transaction with the given hash on the blockchain).

ScriptFailure ScriptError
For pay-to-script outputs: evaluation of the validator script failed.

CardanoLedgerValidationError Text
An error from Cardano.Ledger validation

MaxCollateralInputsExceeded
Balancing failed, it needed more than the maximum number of collateral inputs

class AsValidationError r where

Minimal complete definition

_ValidationError

Methods

_ValidationError :: Prism' r ValidationError

_TxOutRefNotFound :: Prism' r TxIn

_ScriptFailure :: Prism' r ScriptError

_CardanoLedgerValidationError :: Prism' r Text

_MaxCollateralInputsExceeded :: Prism' r ()

data ValidationPhase

Constructors

Phase1

Phase2

type ValidationErrorInPhase = (ValidationPhase, ValidationError)

type ValidationSuccess = Map RdmrPtr ([Text], ExUnits)

Orphan instances

Semigroup (UTxO era)

Instance details

Methods

(<>) :: UTxO era -> UTxO era -> UTxO era

sconcat :: NonEmpty (UTxO era) -> UTxO era

stimes :: Integral b => b -> UTxO era -> UTxO era

Monoid (UTxO era)

Instance details

Methods

mempty :: UTxO era

mappend :: UTxO era -> UTxO era -> UTxO era

mconcat :: [UTxO era] -> UTxO era