Ledger.Orphans

Orphan instances

Data NetworkId

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> NetworkId -> c NetworkId

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c NetworkId

toConstr :: NetworkId -> Constr

dataTypeOf :: NetworkId -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c NetworkId)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c NetworkId)

gmapT :: (forall b. Data b => b -> b) -> NetworkId -> NetworkId

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> NetworkId -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> NetworkId -> r

gmapQ :: (forall d. Data d => d -> u) -> NetworkId -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> NetworkId -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> NetworkId -> m NetworkId

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> NetworkId -> m NetworkId

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> NetworkId -> m NetworkId

Data NetworkMagic

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> NetworkMagic -> c NetworkMagic

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c NetworkMagic

toConstr :: NetworkMagic -> Constr

dataTypeOf :: NetworkMagic -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c NetworkMagic)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c NetworkMagic)

gmapT :: (forall b. Data b => b -> b) -> NetworkMagic -> NetworkMagic

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> NetworkMagic -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> NetworkMagic -> r

gmapQ :: (forall d. Data d => d -> u) -> NetworkMagic -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> NetworkMagic -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> NetworkMagic -> m NetworkMagic

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> NetworkMagic -> m NetworkMagic

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> NetworkMagic -> m NetworkMagic

Generic Lovelace

Instance details

Associated Types

type Rep Lovelace :: Type -> Type

Methods

from :: Lovelace -> Rep Lovelace x

to :: Rep Lovelace x -> Lovelace

Generic Quantity

Instance details

Associated Types

type Rep Quantity :: Type -> Type

Methods

from :: Quantity -> Rep Quantity x

to :: Rep Quantity x -> Quantity

Generic PolicyId

Instance details

Associated Types

type Rep PolicyId :: Type -> Type

Methods

from :: PolicyId -> Rep PolicyId x

to :: Rep PolicyId x -> PolicyId

Generic NetworkId

Instance details

Associated Types

type Rep NetworkId :: Type -> Type

Methods

from :: NetworkId -> Rep NetworkId x

to :: Rep NetworkId x -> NetworkId

Hashable POSIXTime

Instance details

Methods

hashWithSalt :: Int -> POSIXTime -> Int

hash :: POSIXTime -> Int

ToJSON ScriptError

Instance details

Methods

toJSON :: ScriptError -> Value

toEncoding :: ScriptError -> Encoding

toJSONList :: [ScriptError] -> Value

toEncodingList :: [ScriptError] -> Encoding

ToJSON POSIXTime
Custom ToJSON instance which allows to simply convert a POSIXTime value to a JSON number.

Instance details

Methods

toJSON :: POSIXTime -> Value

toEncoding :: POSIXTime -> Encoding

toJSONList :: [POSIXTime] -> Value

toEncodingList :: [POSIXTime] -> Encoding

FromJSON ScriptError

Instance details

Methods

parseJSON :: Value -> Parser ScriptError

parseJSONList :: Value -> Parser [ScriptError]

FromJSON POSIXTime
Custom FromJSON instance which allows to parse a JSON number to a POSIXTime value. The parsed JSON value MUST be an Integer or else the parsing fails.

Instance details

Methods

parseJSON :: Value -> Parser POSIXTime

parseJSONList :: Value -> Parser [POSIXTime]

ByteArrayAccess TxId
ByteArrayAccess instance for signing support

Instance details

Methods

length :: TxId -> Int

withByteArray :: TxId -> (Ptr p -> IO a) -> IO a

copyByteArrayToPtr :: TxId -> Ptr p -> IO ()

ToHttpApiData LedgerBytes

Instance details

Methods

toUrlPiece :: LedgerBytes -> Text

toEncodedUrlPiece :: LedgerBytes -> Builder

toHeader :: LedgerBytes -> ByteString

toQueryParam :: LedgerBytes -> Text

ToHttpApiData PrivateKey

Instance details

Methods

toUrlPiece :: PrivateKey -> Text

toEncodedUrlPiece :: PrivateKey -> Builder

toHeader :: PrivateKey -> ByteString

toQueryParam :: PrivateKey -> Text

FromHttpApiData LedgerBytes

Instance details

Methods

parseUrlPiece :: Text -> Either Text LedgerBytes

parseHeader :: ByteString -> Either Text LedgerBytes

parseQueryParam :: Text -> Either Text LedgerBytes

FromHttpApiData PrivateKey

Instance details

Methods

parseUrlPiece :: Text -> Either Text PrivateKey

parseHeader :: ByteString -> Either Text PrivateKey

parseQueryParam :: Text -> Either Text PrivateKey

Serialise POSIXTime

Instance details

Methods

encode :: POSIXTime -> Encoding

decode :: Decoder s POSIXTime

encodeList :: [POSIXTime] -> Encoding

decodeList :: Decoder s [POSIXTime]

Serialise (AddressInEra BabbageEra)

Instance details

Methods

encode :: AddressInEra BabbageEra -> Encoding

decode :: Decoder s (AddressInEra BabbageEra)

encodeList :: [AddressInEra BabbageEra] -> Encoding

decodeList :: Decoder s [AddressInEra BabbageEra]