Ledger.Slot

Description

Slots and slot ranges.

Documentation

newtype Slot

The slot number. This is a good proxy for time, since on the Cardano blockchain slots pass at a constant rate.

Constructors

Slot

Fields

getSlot :: Integer

type SlotRange = Interval Slot

An Interval of Slots.

width :: SlotRange -> Maybe Integer

Number of Slots covered by the interval, if finite. width (from x) == Nothing.

Orphan instances

Serialise a => Serialise (Interval a)

Instance details

Methods

encode :: Interval a -> Encoding

decode :: Decoder s (Interval a)

encodeList :: [Interval a] -> Encoding

decodeList :: Decoder s [Interval a]

Serialise a => Serialise (Extended a)

Instance details

Methods

encode :: Extended a -> Encoding

decode :: Decoder s (Extended a)

encodeList :: [Extended a] -> Encoding

decodeList :: Decoder s [Extended a]

Serialise a => Serialise (LowerBound a)

Instance details

Methods

encode :: LowerBound a -> Encoding

decode :: Decoder s (LowerBound a)

encodeList :: [LowerBound a] -> Encoding

decodeList :: Decoder s [LowerBound a]

Serialise a => Serialise (UpperBound a)

Instance details

Methods

encode :: UpperBound a -> Encoding

decode :: Decoder s (UpperBound a)

encodeList :: [UpperBound a] -> Encoding

decodeList :: Decoder s [UpperBound a]

Hashable a => Hashable (Interval a)

Instance details

Methods

hashWithSalt :: Int -> Interval a -> Int

hash :: Interval a -> Int

Hashable a => Hashable (Extended a)

Instance details

Methods

hashWithSalt :: Int -> Extended a -> Int

hash :: Extended a -> Int

Hashable a => Hashable (LowerBound a)

Instance details

Methods

hashWithSalt :: Int -> LowerBound a -> Int

hash :: LowerBound a -> Int

Hashable a => Hashable (UpperBound a)

Instance details

Methods

hashWithSalt :: Int -> UpperBound a -> Int

hash :: UpperBound a -> Int

FromJSON a => FromJSON (Interval a)

Instance details

Methods

parseJSON :: Value -> Parser (Interval a)

parseJSONList :: Value -> Parser [Interval a]

FromJSON a => FromJSON (Extended a)

Instance details

Methods

parseJSON :: Value -> Parser (Extended a)

parseJSONList :: Value -> Parser [Extended a]

FromJSON a => FromJSON (LowerBound a)

Instance details

Methods

parseJSON :: Value -> Parser (LowerBound a)

parseJSONList :: Value -> Parser [LowerBound a]

FromJSON a => FromJSON (UpperBound a)

Instance details

Methods

parseJSON :: Value -> Parser (UpperBound a)

parseJSONList :: Value -> Parser [UpperBound a]

ToJSON a => ToJSON (Interval a)

Instance details

Methods

toJSON :: Interval a -> Value

toEncoding :: Interval a -> Encoding

toJSONList :: [Interval a] -> Value

toEncodingList :: [Interval a] -> Encoding

ToJSON a => ToJSON (Extended a)

Instance details

Methods

toJSON :: Extended a -> Value

toEncoding :: Extended a -> Encoding

toJSONList :: [Extended a] -> Value

toEncodingList :: [Extended a] -> Encoding

ToJSON a => ToJSON (LowerBound a)

Instance details

Methods

toJSON :: LowerBound a -> Value

toEncoding :: LowerBound a -> Encoding

toJSONList :: [LowerBound a] -> Value

toEncodingList :: [LowerBound a] -> Encoding

ToJSON a => ToJSON (UpperBound a)

Instance details

Methods

toJSON :: UpperBound a -> Value

toEncoding :: UpperBound a -> Encoding

toJSONList :: [UpperBound a] -> Value

toEncodingList :: [UpperBound a] -> Encoding