Ledger.Tokens

Description

Dealing with tokens

Documentation

The extended UTXO ledger with scripts that Plutus runs on supports native user-defined currencies. Tokens are user-defined currencies with an issuance of exactly one unit. Every token is identified by a Currency.

token :: AssetClass -> Value

A value that contains exactly the token.

outputsWith :: TxInfo -> AssetClass -> [TxOut]

The outputs of the ValidatorCtx that carry a non-zero amount of the currency.

paidTo :: TxInfo -> AssetClass -> Value

The total Value paid by the pending transaction to outputs whose value also includes a non-zero amount of the currency.