Ledger.Tx.CardanoAPI.Internal

Description

Interface to the transaction types from 'cardano-api'

Documentation

newtype CardanoBuildTx

Constructors

CardanoBuildTx

Fields

getCardanoBuildTx :: TxBodyContent BuildTx BabbageEra

data CardanoTx where

Cardano tx from any era.

Constructors

CardanoTx :: IsCardanoEra era => Tx era -> EraInMode era CardanoMode -> CardanoTx

txOutRefs :: CardanoTx -> [(TxOut, TxOutRef)]

unspentOutputsTx :: CardanoTx -> Map TxOutRef TxOut

fromCardanoTxId :: TxId -> TxId

fromCardanoTxIn :: TxIn -> TxOutRef

fromCardanoTxOutToPV1TxInfoTxOut :: TxOut CtxTx era -> TxOut

fromCardanoTxOutToPV1TxInfoTxOut' :: TxOut CtxUTxO era -> TxOut

fromCardanoTxOutToPV2TxInfoTxOut :: TxOut CtxTx era -> TxOut

fromCardanoTxOutToPV2TxInfoTxOut' :: TxOut CtxUTxO era -> TxOut

fromCardanoTxOutDatumHash :: TxOutDatum CtxTx era -> Maybe DatumHash

fromCardanoTxOutDatum :: TxOutDatum CtxTx era -> OutputDatum

fromCardanoTxOutValue :: TxOutValue era -> Value

fromCardanoAddressInEra :: AddressInEra era -> Address

Deprecated: we now use Cardano address internally, if you need a plutus address use toPlutusAddress

fromCardanoAddress :: Address addrtype -> Address

Deprecated: Shouldn't be used as we use Cardano address internally now

fromCardanoAssetId :: AssetId -> AssetClass

fromCardanoAssetName :: AssetName -> TokenName

fromCardanoMintValue :: TxMintValue build era -> Value

fromCardanoValue :: Value -> Value

fromCardanoPolicyId :: PolicyId -> MintingPolicyHash

fromCardanoFee :: TxFee era -> Lovelace

fromCardanoValidityRange :: (TxValidityLowerBound era, TxValidityUpperBound era) -> SlotRange

fromCardanoScriptInEra :: ScriptInEra era -> Maybe (Versioned Script)

fromCardanoPaymentKeyHash :: Hash PaymentKey -> PubKeyHash

Deprecated: Shouldn't be used as we use Cardano address internally now

fromCardanoScriptData :: ScriptData -> BuiltinData

fromCardanoPlutusScript :: HasTypeProxy lang => PlutusScript lang -> Script

fromCardanoScriptInAnyLang :: ScriptInAnyLang -> Maybe (Versioned Script)

fromCardanoReferenceScript :: ReferenceScript BabbageEra -> Maybe (Versioned Script)

fromCardanoLovelace :: Lovelace -> Value

fromCardanoSlotNo :: SlotNo -> Slot

fromTxScriptValidity :: TxScriptValidity era -> Bool

Given a 'C.TxScriptValidity era', if the era supports scripts, return a True or False depending on script validity. If the era does not support scripts, always return True.

toTxScriptValidity :: ShelleyBasedEra era -> Bool -> TxScriptValidity era

scriptDataFromCardanoTxBody :: TxBody era -> (Map DatumHash Datum, Redeemers)

Given a 'C.TxBody from a 'C.Tx era', return the datums and redeemers along with their hashes.

plutusScriptsFromTxBody :: TxBody era -> Map ScriptHash (Versioned Script)

Extract plutus scripts from a Cardano API tx body.

Note that Plutus scripts are only supported in Alonzo era and onwards.

makeTransactionBody :: Maybe (PParams (BabbageEra StandardCrypto)) -> Map RdmrPtr ExUnits -> CardanoBuildTx -> Either ToCardanoError (TxBody BabbageEra)

toCardanoTxIn :: TxOutRef -> Either ToCardanoError TxIn

toCardanoTxOut :: NetworkId -> TxOut -> Either ToCardanoError (TxOut CtxTx BabbageEra)

toCardanoTxOutDatum :: OutputDatum -> Either ToCardanoError (TxOutDatum CtxTx BabbageEra)

toCardanoTxOutDatumHash :: DatumHash -> Either ToCardanoError (TxOutDatum ctx BabbageEra)

toCardanoTxOutDatumHashFromDatum :: Datum -> TxOutDatum ctx BabbageEra

toCardanoTxOutDatumInline :: Datum -> TxOutDatum CtxTx BabbageEra

toCardanoTxOutDatumInTx :: Datum -> TxOutDatum CtxTx BabbageEra

toCardanoTxOutNoDatum :: TxOutDatum CtxTx BabbageEra

toCardanoTxOutValue :: Value -> TxOutValue BabbageEra

toCardanoAddressInEra :: NetworkId -> Address -> Either ToCardanoError (AddressInEra BabbageEra)

toCardanoAssetId :: AssetClass -> Either ToCardanoError AssetId

toCardanoAssetName :: TokenName -> Either ToCardanoError AssetName

toCardanoPolicyId :: MintingPolicyHash -> Either ToCardanoError PolicyId

toCardanoValue :: Value -> Either ToCardanoError Value

toCardanoLovelace :: Value -> Either ToCardanoError Lovelace

toCardanoFee :: Lovelace -> TxFee BabbageEra

adaToCardanoValue :: Ada -> Value

toCardanoValidityRange :: SlotRange -> Either ToCardanoError (TxValidityLowerBound BabbageEra, TxValidityUpperBound BabbageEra)

toCardanoScriptInEra :: Versioned Script -> Either ToCardanoError (ScriptInEra BabbageEra)

toCardanoPaymentKeyHash :: PaymentPubKeyHash -> Either ToCardanoError (Hash PaymentKey)

toCardanoScriptData :: BuiltinData -> ScriptData

toCardanoScriptDataHash :: DatumHash -> Either ToCardanoError (Hash ScriptData)

toCardanoScriptHash :: ValidatorHash -> Either ToCardanoError ScriptHash

toCardanoStakeKeyHash :: PubKeyHash -> Either ToCardanoError (Hash StakeKey)

toCardanoPlutusScript :: SerialiseAsRawBytes plutusScript => AsType plutusScript -> Script -> Either ToCardanoError plutusScript

toCardanoScriptInAnyLang :: Versioned Script -> Either ToCardanoError ScriptInAnyLang

toCardanoReferenceScript :: Maybe (Versioned Script) -> Either ToCardanoError (ReferenceScript BabbageEra)

toCardanoTxId :: TxId -> Either ToCardanoError TxId

data ToCardanoError

Constructors

TxBodyError String
A C.TxBodyError converted to String

DeserialisationError

InvalidValidityRange

ValueNotPureAda

OutputHasZeroAda

StakingPointersNotSupported

SimpleScriptsNotSupportedToCardano

MissingInputValidator

MissingDatum

MissingMintingPolicy

ScriptPurposeNotSupported ScriptTag

MissingMintingPolicyRedeemer

MissingStakeValidator

UnsupportedPlutusVersion Language

Tag String ToCardanoError

data FromCardanoError

Constructors

SimpleScriptsNotSupported

deserialiseFromRawBytes :: SerialiseAsRawBytes t => AsType t -> ByteString -> Either ToCardanoError t

zeroExecutionUnits :: ExecutionUnits

tag :: String -> Either ToCardanoError t -> Either ToCardanoError t

withIsCardanoEra :: EraInMode era CardanoMode -> (IsCardanoEra era => r) -> r

Run code that needs an IsCardanoEra constraint while you only have an EraInMode value.

Orphan instances

ToJSON ScriptData

Instance details

Methods

toJSON :: ScriptData -> Value

toEncoding :: ScriptData -> Encoding

toJSONList :: [ScriptData] -> Value

toEncodingList :: [ScriptData] -> Encoding

ToJSONKey PolicyId

Instance details

Methods

toJSONKey :: ToJSONKeyFunction PolicyId

toJSONKeyList :: ToJSONKeyFunction [PolicyId]

FromJSON ScriptData

Instance details

Methods

parseJSON :: Value -> Parser ScriptData

parseJSONList :: Value -> Parser [ScriptData]

FromJSONKey PolicyId

Instance details

Methods

fromJSONKey :: FromJSONKeyFunction PolicyId

fromJSONKeyList :: FromJSONKeyFunction [PolicyId]

Generic (PlutusScriptOrReferenceInput lang)

Instance details

Associated Types

type Rep (PlutusScriptOrReferenceInput lang) :: Type -> Type

Methods

from :: PlutusScriptOrReferenceInput lang -> Rep (PlutusScriptOrReferenceInput lang) x

to :: Rep (PlutusScriptOrReferenceInput lang) x -> PlutusScriptOrReferenceInput lang

ToJSON (TxScriptValidity BabbageEra)

Instance details

Methods

toJSON :: TxScriptValidity BabbageEra -> Value

toEncoding :: TxScriptValidity BabbageEra -> Encoding

toJSONList :: [TxScriptValidity BabbageEra] -> Value

toEncodingList :: [TxScriptValidity BabbageEra] -> Encoding

ToJSON (TxInsCollateral BabbageEra)

Instance details

Methods

toJSON :: TxInsCollateral BabbageEra -> Value

toEncoding :: TxInsCollateral BabbageEra -> Encoding

toJSONList :: [TxInsCollateral BabbageEra] -> Value

toEncodingList :: [TxInsCollateral BabbageEra] -> Encoding

ToJSON (TxTotalCollateral BabbageEra)

Instance details

Methods

toJSON :: TxTotalCollateral BabbageEra -> Value

toEncoding :: TxTotalCollateral BabbageEra -> Encoding

toJSONList :: [TxTotalCollateral BabbageEra] -> Value

toEncodingList :: [TxTotalCollateral BabbageEra] -> Encoding

ToJSON (TxFee BabbageEra)

Instance details

Methods

toJSON :: TxFee BabbageEra -> Value

toEncoding :: TxFee BabbageEra -> Encoding

toJSONList :: [TxFee BabbageEra] -> Value

toEncodingList :: [TxFee BabbageEra] -> Encoding

ToJSON (TxValidityUpperBound BabbageEra)

Instance details

Methods

toJSON :: TxValidityUpperBound BabbageEra -> Value

toEncoding :: TxValidityUpperBound BabbageEra -> Encoding

toJSONList :: [TxValidityUpperBound BabbageEra] -> Value

toEncodingList :: [TxValidityUpperBound BabbageEra] -> Encoding

ToJSON (TxValidityLowerBound BabbageEra)

Instance details

Methods

toJSON :: TxValidityLowerBound BabbageEra -> Value

toEncoding :: TxValidityLowerBound BabbageEra -> Encoding

toJSONList :: [TxValidityLowerBound BabbageEra] -> Value

toEncodingList :: [TxValidityLowerBound BabbageEra] -> Encoding

ToJSON (TxMetadataInEra BabbageEra)

Instance details

Methods

toJSON :: TxMetadataInEra BabbageEra -> Value

toEncoding :: TxMetadataInEra BabbageEra -> Encoding

toJSONList :: [TxMetadataInEra BabbageEra] -> Value

toEncodingList :: [TxMetadataInEra BabbageEra] -> Encoding

ToJSON (TxAuxScripts BabbageEra)

Instance details

Methods

toJSON :: TxAuxScripts BabbageEra -> Value

toEncoding :: TxAuxScripts BabbageEra -> Encoding

toJSONList :: [TxAuxScripts BabbageEra] -> Value

toEncodingList :: [TxAuxScripts BabbageEra] -> Encoding

ToJSON (TxExtraKeyWitnesses BabbageEra)

Instance details

Methods

toJSON :: TxExtraKeyWitnesses BabbageEra -> Value

toEncoding :: TxExtraKeyWitnesses BabbageEra -> Encoding

toJSONList :: [TxExtraKeyWitnesses BabbageEra] -> Value

toEncodingList :: [TxExtraKeyWitnesses BabbageEra] -> Encoding

ToJSON (TxUpdateProposal BabbageEra)

Instance details

Methods

toJSON :: TxUpdateProposal BabbageEra -> Value

toEncoding :: TxUpdateProposal BabbageEra -> Encoding

toJSONList :: [TxUpdateProposal BabbageEra] -> Value

toEncodingList :: [TxUpdateProposal BabbageEra] -> Encoding

ToJSON (ScriptInEra BabbageEra)

Instance details

Methods

toJSON :: ScriptInEra BabbageEra -> Value

toEncoding :: ScriptInEra BabbageEra -> Encoding

toJSONList :: [ScriptInEra BabbageEra] -> Value

toEncodingList :: [ScriptInEra BabbageEra] -> Encoding

ToJSON (PlutusScriptOrReferenceInput PlutusScriptV1)

Instance details

Methods

toJSON :: PlutusScriptOrReferenceInput PlutusScriptV1 -> Value

toEncoding :: PlutusScriptOrReferenceInput PlutusScriptV1 -> Encoding

toJSONList :: [PlutusScriptOrReferenceInput PlutusScriptV1] -> Value

toEncodingList :: [PlutusScriptOrReferenceInput PlutusScriptV1] -> Encoding

ToJSON (PlutusScriptOrReferenceInput PlutusScriptV2)

Instance details

Methods

toJSON :: PlutusScriptOrReferenceInput PlutusScriptV2 -> Value

toEncoding :: PlutusScriptOrReferenceInput PlutusScriptV2 -> Encoding

toJSONList :: [PlutusScriptOrReferenceInput PlutusScriptV2] -> Value

toEncodingList :: [PlutusScriptOrReferenceInput PlutusScriptV2] -> Encoding

ToJSON (ScriptDatum WitCtxTxIn)

Instance details

Methods

toJSON :: ScriptDatum WitCtxTxIn -> Value

toEncoding :: ScriptDatum WitCtxTxIn -> Encoding

toJSONList :: [ScriptDatum WitCtxTxIn] -> Value

toEncodingList :: [ScriptDatum WitCtxTxIn] -> Encoding

ToJSON (ScriptDatum WitCtxMint)

Instance details

Methods

toJSON :: ScriptDatum WitCtxMint -> Value

toEncoding :: ScriptDatum WitCtxMint -> Encoding

toJSONList :: [ScriptDatum WitCtxMint] -> Value

toEncodingList :: [ScriptDatum WitCtxMint] -> Encoding

ToJSON (ScriptDatum WitCtxStake)

Instance details

Methods

toJSON :: ScriptDatum WitCtxStake -> Value

toEncoding :: ScriptDatum WitCtxStake -> Encoding

toJSONList :: [ScriptDatum WitCtxStake] -> Value

toEncodingList :: [ScriptDatum WitCtxStake] -> Encoding

ToJSON (PlutusScript PlutusScriptV1)

Instance details

Methods

toJSON :: PlutusScript PlutusScriptV1 -> Value

toEncoding :: PlutusScript PlutusScriptV1 -> Encoding

toJSONList :: [PlutusScript PlutusScriptV1] -> Value

toEncodingList :: [PlutusScript PlutusScriptV1] -> Encoding

ToJSON (PlutusScript PlutusScriptV2)

Instance details

Methods

toJSON :: PlutusScript PlutusScriptV2 -> Value

toEncoding :: PlutusScript PlutusScriptV2 -> Encoding

toJSONList :: [PlutusScript PlutusScriptV2] -> Value

toEncodingList :: [PlutusScript PlutusScriptV2] -> Encoding

FromJSON (TxScriptValidity BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxScriptValidity BabbageEra)

parseJSONList :: Value -> Parser [TxScriptValidity BabbageEra]

FromJSON (TxInsCollateral BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxInsCollateral BabbageEra)

parseJSONList :: Value -> Parser [TxInsCollateral BabbageEra]

FromJSON (TxTotalCollateral BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxTotalCollateral BabbageEra)

parseJSONList :: Value -> Parser [TxTotalCollateral BabbageEra]

FromJSON (TxFee BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxFee BabbageEra)

parseJSONList :: Value -> Parser [TxFee BabbageEra]

FromJSON (TxValidityUpperBound BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxValidityUpperBound BabbageEra)

parseJSONList :: Value -> Parser [TxValidityUpperBound BabbageEra]

FromJSON (TxValidityLowerBound BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxValidityLowerBound BabbageEra)

parseJSONList :: Value -> Parser [TxValidityLowerBound BabbageEra]

FromJSON (TxMetadataInEra BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxMetadataInEra BabbageEra)

parseJSONList :: Value -> Parser [TxMetadataInEra BabbageEra]

FromJSON (TxAuxScripts BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxAuxScripts BabbageEra)

parseJSONList :: Value -> Parser [TxAuxScripts BabbageEra]

FromJSON (TxExtraKeyWitnesses BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxExtraKeyWitnesses BabbageEra)

parseJSONList :: Value -> Parser [TxExtraKeyWitnesses BabbageEra]

FromJSON (TxUpdateProposal BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxUpdateProposal BabbageEra)

parseJSONList :: Value -> Parser [TxUpdateProposal BabbageEra]

FromJSON (ScriptInEra BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (ScriptInEra BabbageEra)

parseJSONList :: Value -> Parser [ScriptInEra BabbageEra]

FromJSON (PlutusScriptOrReferenceInput PlutusScriptV1)

Instance details

Methods

parseJSON :: Value -> Parser (PlutusScriptOrReferenceInput PlutusScriptV1)

parseJSONList :: Value -> Parser [PlutusScriptOrReferenceInput PlutusScriptV1]

FromJSON (PlutusScriptOrReferenceInput PlutusScriptV2)

Instance details

Methods

parseJSON :: Value -> Parser (PlutusScriptOrReferenceInput PlutusScriptV2)

parseJSONList :: Value -> Parser [PlutusScriptOrReferenceInput PlutusScriptV2]

FromJSON (ScriptDatum WitCtxTxIn)

Instance details

Methods

parseJSON :: Value -> Parser (ScriptDatum WitCtxTxIn)

parseJSONList :: Value -> Parser [ScriptDatum WitCtxTxIn]

FromJSON (ScriptDatum WitCtxMint)

Instance details

Methods

parseJSON :: Value -> Parser (ScriptDatum WitCtxMint)

parseJSONList :: Value -> Parser [ScriptDatum WitCtxMint]

FromJSON (ScriptDatum WitCtxStake)

Instance details

Methods

parseJSON :: Value -> Parser (ScriptDatum WitCtxStake)

parseJSONList :: Value -> Parser [ScriptDatum WitCtxStake]

FromJSON (PlutusScript PlutusScriptV1)

Instance details

Methods

parseJSON :: Value -> Parser (PlutusScript PlutusScriptV1)

parseJSONList :: Value -> Parser [PlutusScript PlutusScriptV1]

FromJSON (PlutusScript PlutusScriptV2)

Instance details

Methods

parseJSON :: Value -> Parser (PlutusScript PlutusScriptV2)

parseJSONList :: Value -> Parser [PlutusScript PlutusScriptV2]

Generic (TxBodyContent BuildTx BabbageEra)

Instance details

Associated Types

type Rep (TxBodyContent BuildTx BabbageEra) :: Type -> Type

Methods

from :: TxBodyContent BuildTx BabbageEra -> Rep (TxBodyContent BuildTx BabbageEra) x

to :: Rep (TxBodyContent BuildTx BabbageEra) x -> TxBodyContent BuildTx BabbageEra

ToJSON w => ToJSON (BuildTxWith BuildTx w)

Instance details

Methods

toJSON :: BuildTxWith BuildTx w -> Value

toEncoding :: BuildTxWith BuildTx w -> Encoding

toJSONList :: [BuildTxWith BuildTx w] -> Value

toEncodingList :: [BuildTxWith BuildTx w] -> Encoding

ToJSON (TxInsReference BuildTx BabbageEra)

Instance details

Methods

toJSON :: TxInsReference BuildTx BabbageEra -> Value

toEncoding :: TxInsReference BuildTx BabbageEra -> Encoding

toJSONList :: [TxInsReference BuildTx BabbageEra] -> Value

toEncodingList :: [TxInsReference BuildTx BabbageEra] -> Encoding

ToJSON (TxReturnCollateral ctx BabbageEra)

Instance details

Methods

toJSON :: TxReturnCollateral ctx BabbageEra -> Value

toEncoding :: TxReturnCollateral ctx BabbageEra -> Encoding

toJSONList :: [TxReturnCollateral ctx BabbageEra] -> Value

toEncodingList :: [TxReturnCollateral ctx BabbageEra] -> Encoding

ToJSON (TxWithdrawals BuildTx BabbageEra)

Instance details

Methods

toJSON :: TxWithdrawals BuildTx BabbageEra -> Value

toEncoding :: TxWithdrawals BuildTx BabbageEra -> Encoding

toJSONList :: [TxWithdrawals BuildTx BabbageEra] -> Value

toEncodingList :: [TxWithdrawals BuildTx BabbageEra] -> Encoding

ToJSON (TxCertificates BuildTx BabbageEra)

Instance details

Methods

toJSON :: TxCertificates BuildTx BabbageEra -> Value

toEncoding :: TxCertificates BuildTx BabbageEra -> Encoding

toJSONList :: [TxCertificates BuildTx BabbageEra] -> Value

toEncodingList :: [TxCertificates BuildTx BabbageEra] -> Encoding

ToJSON (TxMintValue BuildTx BabbageEra)

Instance details

Methods

toJSON :: TxMintValue BuildTx BabbageEra -> Value

toEncoding :: TxMintValue BuildTx BabbageEra -> Encoding

toJSONList :: [TxMintValue BuildTx BabbageEra] -> Value

toEncodingList :: [TxMintValue BuildTx BabbageEra] -> Encoding

ToJSON (TxBodyContent BuildTx BabbageEra)

Instance details

Methods

toJSON :: TxBodyContent BuildTx BabbageEra -> Value

toEncoding :: TxBodyContent BuildTx BabbageEra -> Encoding

toJSONList :: [TxBodyContent BuildTx BabbageEra] -> Value

toEncodingList :: [TxBodyContent BuildTx BabbageEra] -> Encoding

ToJSON (ScriptDatum ctx) => ToJSON (ScriptWitness ctx BabbageEra)

Instance details

Methods

toJSON :: ScriptWitness ctx BabbageEra -> Value

toEncoding :: ScriptWitness ctx BabbageEra -> Encoding

toJSONList :: [ScriptWitness ctx BabbageEra] -> Value

toEncodingList :: [ScriptWitness ctx BabbageEra] -> Encoding

ToJSON (Witness WitCtxTxIn BabbageEra)

Instance details

Methods

toJSON :: Witness WitCtxTxIn BabbageEra -> Value

toEncoding :: Witness WitCtxTxIn BabbageEra -> Encoding

toJSONList :: [Witness WitCtxTxIn BabbageEra] -> Value

toEncodingList :: [Witness WitCtxTxIn BabbageEra] -> Encoding

ToJSON (Witness WitCtxMint BabbageEra)

Instance details

Methods

toJSON :: Witness WitCtxMint BabbageEra -> Value

toEncoding :: Witness WitCtxMint BabbageEra -> Encoding

toJSONList :: [Witness WitCtxMint BabbageEra] -> Value

toEncodingList :: [Witness WitCtxMint BabbageEra] -> Encoding

ToJSON (Witness WitCtxStake BabbageEra)

Instance details

Methods

toJSON :: Witness WitCtxStake BabbageEra -> Value

toEncoding :: Witness WitCtxStake BabbageEra -> Encoding

toJSONList :: [Witness WitCtxStake BabbageEra] -> Value

toEncodingList :: [Witness WitCtxStake BabbageEra] -> Encoding

FromJSON w => FromJSON (BuildTxWith BuildTx w)

Instance details

Methods

parseJSON :: Value -> Parser (BuildTxWith BuildTx w)

parseJSONList :: Value -> Parser [BuildTxWith BuildTx w]

FromJSON (TxInsReference BuildTx BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxInsReference BuildTx BabbageEra)

parseJSONList :: Value -> Parser [TxInsReference BuildTx BabbageEra]

FromJSON (TxReturnCollateral CtxTx BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxReturnCollateral CtxTx BabbageEra)

parseJSONList :: Value -> Parser [TxReturnCollateral CtxTx BabbageEra]

FromJSON (TxWithdrawals BuildTx BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxWithdrawals BuildTx BabbageEra)

parseJSONList :: Value -> Parser [TxWithdrawals BuildTx BabbageEra]

FromJSON (TxCertificates BuildTx BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxCertificates BuildTx BabbageEra)

parseJSONList :: Value -> Parser [TxCertificates BuildTx BabbageEra]

FromJSON (TxMintValue BuildTx BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxMintValue BuildTx BabbageEra)

parseJSONList :: Value -> Parser [TxMintValue BuildTx BabbageEra]

FromJSON (TxBodyContent BuildTx BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (TxBodyContent BuildTx BabbageEra)

parseJSONList :: Value -> Parser [TxBodyContent BuildTx BabbageEra]

FromJSON (ScriptDatum ctx) => FromJSON (ScriptWitness ctx BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (ScriptWitness ctx BabbageEra)

parseJSONList :: Value -> Parser [ScriptWitness ctx BabbageEra]

FromJSON (Witness WitCtxTxIn BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (Witness WitCtxTxIn BabbageEra)

parseJSONList :: Value -> Parser [Witness WitCtxTxIn BabbageEra]

FromJSON (Witness WitCtxMint BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (Witness WitCtxMint BabbageEra)

parseJSONList :: Value -> Parser [Witness WitCtxMint BabbageEra]

FromJSON (Witness WitCtxStake BabbageEra)

Instance details

Methods

parseJSON :: Value -> Parser (Witness WitCtxStake BabbageEra)

parseJSONList :: Value -> Parser [Witness WitCtxStake BabbageEra]