Ledger.Tx.Internal

Documentation

type ReferenceScript = ReferenceScript BabbageEra

type MintingWitnessesMap = Map MintingPolicyHash (Redeemer, Maybe (Versioned TxOutRef))

type ScriptsMap = Map ScriptHash (Versioned Script)

newtype TxOut

Constructors

TxOut

Fields

getTxOut :: TxOut CtxTx BabbageEra

data Certificate

Constructors

Certificate

Fields

certificateDcert :: DCert

certificateRedeemer :: Maybe Redeemer

redeemer for script credential

data Withdrawal

Stake withdrawal, if applicable the script should be included in txScripts.

Constructors

Withdrawal

Fields

withdrawalCredential :: Credential

staking credential

withdrawalAmount :: Integer

amount of withdrawal in Lovelace, must withdraw all eligible amount

withdrawalRedeemer :: Maybe Redeemer

redeemer for script credential

cardanoTxOutValue :: TxOut ctx era -> Value

txOutValue :: TxOut -> Value

outValue :: Lens TxOut TxOut Value (TxOutValue BabbageEra)

outValue' :: Lens' TxOut (TxOutValue BabbageEra)

toSizedTxOut :: TxOut -> Sized (TxOut StandardBabbage)

toCtxUTxOTxOut :: TxOut -> TxOut CtxUTxO BabbageEra

txOutDatumHash :: TxOut -> Maybe DatumHash

Get a hash from the stored TxOutDatum (either directly or by hashing the inlined datum)

cardanoTxOutDatumHash :: TxOutDatum CtxUTxO BabbageEra -> Maybe (Hash ScriptData)

txOutPubKey :: TxOut -> Maybe PubKeyHash

txOutAddress :: TxOut -> CardanoAddress

outAddress :: Lens' TxOut (AddressInEra BabbageEra)

outDatumHash :: Lens TxOut TxOut (Maybe DatumHash) (TxOutDatum CtxTx BabbageEra)

txOutReferenceScript :: TxOut -> ReferenceScript

outReferenceScript :: Lens' TxOut ReferenceScript

lookupScript :: ScriptsMap -> ScriptHash -> Maybe (Versioned Script)

lookupValidator :: ScriptsMap -> ValidatorHash -> Maybe (Versioned Validator)

lookupMintingPolicy :: ScriptsMap -> MintingPolicyHash -> Maybe (Versioned MintingPolicy)

lookupStakeValidator :: ScriptsMap -> StakeValidatorHash -> Maybe (Versioned StakeValidator)

emptyTxBodyContent :: TxBodyContent BuildTx BabbageEra

data Language

Non-Native Script language. This is an Enumerated type. This is expected to be an open type. We will add new Constuctors to this type as additional Non-Native scripting language as are added. We use an enumerated type for two reasons. 1) We can write total functions by case analysis over the constructors 2) We will use DataKinds to make some datatypes indexed by Language For now, the only Non-Native Scriting language is Plutus We might add new languages in the futures.

Note that the the serialization of Language depends on the ordering.

Constructors

PlutusV1

PlutusV2

newtype TxOut

Constructors

TxOut

Fields

getTxOut :: TxOut CtxTx BabbageEra

data TxOutRef

A reference to a transaction output. This is a pair of a transaction reference, and an index indicating which of the outputs of that transaction we are referring to.

Constructors

TxOutRef

Fields

txOutRefId :: TxId

txOutRefIdx :: Integer

Index into the referenced transaction's outputs

data Versioned script

A script of some kind with its Plutus language version

Constructors

Versioned

Fields

unversioned :: script

version :: Language