Ledger.Typed.Scripts

Documentation

data Language

Non-Native Script language. This is an Enumerated type. This is expected to be an open type. We will add new Constuctors to this type as additional Non-Native scripting language as are added. We use an enumerated type for two reasons. 1) We can write total functions by case analysis over the constructors 2) We will use DataKinds to make some datatypes indexed by Language For now, the only Non-Native Scriting language is Plutus We might add new languages in the futures.

Note that the the serialization of Language depends on the ordering.

Constructors

PlutusV1

PlutusV2

forwardingMintingPolicyHash :: TypedValidator a -> MintingPolicyHash

Hash of the minting policy that forwards all checks to the instance's validator

vForwardingMintingPolicy :: TypedValidator a -> Versioned MintingPolicy

The minting policy that forwards all checks to the instance's validator

forwardingMintingPolicy :: TypedValidator a -> MintingPolicy

The unversioned minting policy that forwards all checks to the instance's validator

generalise :: TypedValidator a -> TypedValidator Any

Generalise the typed validator to one that works with the Data type.

vValidatorScript :: TypedValidator a -> Versioned Validator

The validator script itself.

validatorScript :: TypedValidator a -> Validator

The unversioned validator script itself.

validatorCardanoAddress :: NetworkId -> TypedValidator a -> AddressInEra BabbageEra

The address of the validator.

validatorAddress :: TypedValidator a -> Address

The address of the validator.

validatorHash :: TypedValidator a -> ValidatorHash

The hash of the validator.

type UntypedValidator = BuiltinData -> BuiltinData -> BuiltinData -> ()

type UntypedMintingPolicy = BuiltinData -> BuiltinData -> ()

type UntypedStakeValidator = BuiltinData -> BuiltinData -> ()

data Any

type family DatumType a

The type of the data of this connection type.

type family RedeemerType a

The type of the redeemers of this connection type.

class ValidatorTypes a

A class that associates a type standing for a connection type with two types, the type of the redeemer and the data script for that connection type.

Associated Types

type RedeemerType a

The type of the redeemers of this connection type.

type RedeemerType a = ()

type DatumType a

The type of the data of this connection type.

type DatumType a = ()

data TypedValidator a

A typed validator script with its ValidatorScript and Address.

Constructors

TypedValidator

Fields

tvValidator :: Versioned Validator

tvValidatorHash :: ValidatorHash

tvForwardingMPS :: Versioned MintingPolicy

tvForwardingMPSHash :: MintingPolicyHash

The hash of the minting policy that checks whether the validator is run in this transaction

class UnsafeFromData sc => IsScriptContext sc where

Minimal complete definition

Nothing

Methods

mkUntypedValidator :: (UnsafeFromData d, UnsafeFromData r) => (d -> r -> sc -> Bool) -> UntypedValidator

Converts a custom datum and redeemer from a validator function to an untyped validator function. See Note [Scripts returning Bool].

Here's an example of how this function can be used:

  import PlutusTx qualified
  import Plutus.V2.Ledger.Scripts qualified as Plutus
  import Plutus.Script.Utils.V2.Scripts (mkUntypedValidator)

  newtype MyCustomDatum = MyCustomDatum Integer
  PlutusTx.unstableMakeIsData ''MyCustomDatum
  newtype MyCustomRedeemer = MyCustomRedeemer Integer
  PlutusTx.unstableMakeIsData ''MyCustomRedeemer

  mkValidator :: MyCustomDatum -> MyCustomRedeemer -> Plutus.ScriptContext -> Bool
  mkValidator _ _ _ = True

  validator :: Plutus.Validator
  validator = Plutus.mkValidatorScript
      $$(PlutusTx.compile [|| wrap ||])
   where
      wrap = mkUntypedValidator mkValidator

Here's an example using a parameterized validator:

  import PlutusTx qualified
  import Plutus.V2.Ledger.Scripts qualified as Plutus
  import Plutus.Script.Utils.V2.Scripts (mkUntypedValidator)

  newtype MyCustomDatum = MyCustomDatum Integer
  PlutusTx.unstableMakeIsData ''MyCustomDatum
  newtype MyCustomRedeemer = MyCustomRedeemer Integer
  PlutusTx.unstableMakeIsData ''MyCustomRedeemer

  mkValidator :: Int -> MyCustomDatum -> MyCustomRedeemer -> Plutus.ScriptContext -> Bool
  mkValidator _ _ _ _ = True

  validator :: Int -> Plutus.Validator
  validator i = Plutus.mkValidatorScript
      $$(PlutusTx.compile [|| wrap . mkValidator ||]) applyCode PlutusTx.liftCode i
   where
      wrap = mkUntypedValidator

For debugging purpose, it may be of interest to know that in the default implementation, the parameters are decoded in the order they appear (data, redeemer and then script context). A log trace is generated after each successfully decoded parameter. Thus, if a parameter can't be decoded, the culprit is the first parameter in the list that doesn't appear as successfully decoded in the log trace.

mkUntypedStakeValidator :: UnsafeFromData r => (r -> sc -> Bool) -> UntypedStakeValidator

Converts a custom redeemer from a stake validator function to an untyped stake validator function. See Note [Scripts returning Bool].

Here's an example of how this function can be used:

  import PlutusTx qualified
  import Plutus.V1.Ledger.Scripts qualified as Plutus
  import Plutus.Script.Utils.V1.Scripts (mkUntypedStakeValidator)

  newtype MyCustomRedeemer = MyCustomRedeemer Integer
  PlutusTx.unstableMakeIsData ''MyCustomRedeemer

  mkStakeValidator :: MyCustomRedeemer -> ScriptContext -> Bool
  mkStakeValidator _ _ = True

  validator :: Plutus.Validator
  validator = Plutus.mkStakeValidatorScript
      $$(PlutusTx.compile [|| wrap ||])
   where
      wrap = mkUntypedStakeValidator mkStakeValidator

For debugging purpose, it may be of interest to know that in the default implementation, the parameters are decoded in the order they appear (redeemer and then script context). A log trace is generated after each successfully decoded parameter. Thus, if a parameter can't be decoded, the culprit is the first parameter in the list that doesn't appear as successfully decoded in the log trace.

mkUntypedMintingPolicy :: UnsafeFromData r => (r -> sc -> Bool) -> UntypedMintingPolicy

Converts a custom redeemer from a minting policy function to an untyped minting policy function. See Note [Scripts returning Bool].

Here's an example of how this function can be used:

  import PlutusTx qualified
  import Plutus.V1.Ledger.Scripts qualified as Plutus
  import Plutus.Script.Utils.V1.Scripts (mkUntypedMintingPolicy)

  newtype MyCustomRedeemer = MyCustomRedeemer Integer
  PlutusTx.unstableMakeIsData ''MyCustomRedeemer

  mkMintingPolicy :: MyCustomRedeemer -> ScriptContext -> Bool
  mkMintingPolicy _ _ = True

  validator :: Plutus.Validator
  validator = Plutus.mkMintingPolicyScript
      $$(PlutusTx.compile [|| wrap ||])
   where
      wrap = mkUntypedMintingPolicy mkMintingPolicy

For debugging purpose, it may be of interest to know that in the default implementation, the parameters are decoded in the order they appear (redeemer and then script context). A log trace is generated after each successfully decoded parameter. Thus, if a parameter can't be decoded, the culprit is the first parameter in the list that doesn't appear as successfully decoded in the log trace.

type ScriptContextV1 = ScriptContext

type ScriptContextV2 = ScriptContext

data Versioned script

A script of some kind with its Plutus language version

Constructors

Versioned

Fields

unversioned :: script

version :: Language

data MintingPolicy

MintingPolicy is a wrapper around Scripts which are used as validators for minting constraints.

data Validator

Validator is a wrapper around Scripts which are used as validators in transaction outputs.

data ConnectionError

An error we can get while trying to type an existing transaction part.

Constructors

WrongValidatorAddress Address Address

WrongOutType WrongOutTypeError

WrongValidatorType String

WrongRedeemerType BuiltinData

WrongDatumType BuiltinData

NoDatum TxOutRef DatumHash

UnknownRef TxOutRef

mkForwardingMintingPolicy :: Versioned Validator -> Versioned MintingPolicy

unsafeMkTypedValidator :: Versioned Validator -> TypedValidator Any

Make a TypedValidator (with no type constraints) from an untyped Validator script.

type ValidatorType a = DatumType a -> RedeemerType a -> ScriptContext -> Bool

The type of validators for the given connection type.

mkTypedValidator

Arguments

:: CompiledCode (ValidatorType a)
Validator script (compiled)

-> CompiledCode (ValidatorType a -> UntypedValidator)
A wrapper for the compiled validator

-> TypedValidator a

Make a TypedValidator from the CompiledCode of a validator script and its wrapper.

mkTypedValidatorParam

Arguments

:: forall a param. Lift DefaultUni param

=> CompiledCode (param -> ValidatorType a)
Validator script (compiled)

-> CompiledCode (ValidatorType a -> UntypedValidator)
A wrapper for the compiled validator

-> param
The extra paramater for the validator script

-> TypedValidator a

Make a TypedValidator from the CompiledCode of a parameterized validator script and its wrapper.