Prettyprinter.Extras

Description

Copied from plutus-ledger-api because not exported

Documentation

newtype PrettyShow a

Newtype wrapper for deriving Pretty via a Show instance

Constructors

PrettyShow

Fields

unPrettyShow :: a

class Pretty a where

Minimal complete definition

pretty

Methods

pretty :: a -> Doc ann

prettyList :: [a] -> Doc ann

newtype PrettyFoldable f a

Newtype wrapper for deriving Pretty for a Foldable container by calling toList.

Constructors

PrettyFoldable

Fields

unPrettyFoldable :: f a

newtype Tagged (s :: k) b

Constructors

Tagged b

Orphan instances

(KnownSymbol a, Pretty b) => Pretty (Tagged a b)

Instance details

Methods

pretty :: Tagged a b -> Doc ann

prettyList :: [Tagged a b] -> Doc ann