Ledger.AddressMap

Description

AddressMaps and functions for working on them.

AddressMaps are used to represent the limited knowledge about the state of the ledger that the wallet retains. Rather than keeping the entire ledger (which can be very large) the wallet only tracks the UTxOs at particular addresses.

Documentation

newtype AddressMap

A map of Addresses and their unspent outputs.

Constructors

AddressMap

Fields

getAddressMap :: Map CardanoAddress UtxoMap

type UtxoMap = Map TxOutRef (CardanoTx, TxOut)

addAddress :: CardanoAddress -> AddressMap -> AddressMap

Add an address with no unspent outputs to a map. If the address already exists, do nothing.

addAddresses :: [CardanoAddress] -> AddressMap -> AddressMap

Add a list of Addresses with no unspent outputs to the map.

filterRefs :: (TxOutRef -> (CardanoTx, TxOut) -> Bool) -> AddressMap -> AddressMap

Filter the transaction output references in the map

fundsAt :: CardanoAddress -> Lens' AddressMap UtxoMap

Get the funds available at a particular address.

values :: AddressMap -> Map CardanoAddress Value

The total value of unspent outputs (which the map knows about) at an address.

traverseWithKey :: Applicative f => (CardanoAddress -> Map TxOutRef (CardanoTx, TxOut) -> f (Map TxOutRef (CardanoTx, TxOut))) -> AddressMap -> f AddressMap

Walk through the address map, applying an effectful function to each entry.

singleton :: (CardanoAddress, TxOutRef, CardanoTx, TxOut) -> AddressMap

An address map with a single unspent transaction output.

fromTxOutputs :: OnChainTx -> AddressMap

Create an AddressMap with the unspent outputs of a single transaction.

knownAddresses :: AddressMap -> Map TxOutRef CardanoAddress

Create a map of unspent transaction outputs to their addresses (the "inverse" of an AddressMap, without the values)

updateAddresses :: OnChainTx -> AddressMap -> AddressMap

Update an AddressMap with the inputs and outputs of a new transaction. updateAddresses does not add or remove any keys from the map.

updateAllAddresses :: OnChainTx -> AddressMap -> AddressMap

Update an AddressMap with the inputs and outputs of a new transaction, including all addresses in the transaction.

restrict :: AddressMap -> Set CardanoAddress -> AddressMap

Restrict an AddressMap to a set of addresses.

addressesTouched :: AddressMap -> OnChainTx -> Set CardanoAddress

Get the set of all addresses that the transaction spends outputs from or produces outputs to

outRefMap :: AddressMap -> Map TxOutRef (CardanoTx, TxOut)

outputsMapFromTxForAddress :: CardanoAddress -> OnChainTx -> Map TxOutRef (CardanoTx, TxOut)

fromChain :: Blockchain -> AddressMap

The unspent transaction outputs of the ledger as a whole.