Ledger.Blockchain

Documentation

data OnChainTx

A transaction on the blockchain. Invalid transactions are still put on the chain to be able to collect fees.

Constructors

Invalid CardanoTx

Valid CardanoTx

_Valid :: Prism' OnChainTx CardanoTx

_Invalid :: Prism' OnChainTx CardanoTx

type Block = [OnChainTx]

A block on the blockchain. This is just a list of transactions following on from the chain so far.

newtype BlockId

Block identifier (usually a hash)

Constructors

BlockId

Fields

getBlockId :: ByteString

type Blockchain = [Block]

A blockchain, which is just a list of blocks, starting with the newest.

newtype Context

Information about the state of the blockchain and about the transaction that is currently being validated, represented as a value in Data.

Constructors

Context BuiltinData

eitherTx :: (CardanoTx -> r) -> (CardanoTx -> r) -> OnChainTx -> r

unOnChain :: OnChainTx -> CardanoTx

onChainTxIsValid :: OnChainTx -> Bool

consumableInputs :: OnChainTx -> [TxIn]

Outputs consumed from the UTXO set by the OnChainTx

outputsProduced :: OnChainTx -> Map TxOutRef TxOut

Outputs added to the UTXO set by the OnChainTx

transaction :: Blockchain -> TxId -> Maybe OnChainTx

Lookup a transaction in a Blockchain by its id.

out :: Blockchain -> TxOutRef -> Maybe TxOut

Determine the unspent output that an input refers to

value :: Blockchain -> TxOutRef -> Maybe Value

Determine the unspent value that a transaction output refers to.

unspentOutputsTx :: Tx -> Map TxOutRef TxOut

The unspent outputs of a transaction.

spentOutputs :: Tx -> [TxOutRef]

The transaction output references consumed by a transaction.

unspentOutputs :: Blockchain -> Map TxOutRef TxOut

The unspent transaction outputs of the ledger as a whole.

datumTxo :: Blockchain -> TxOutRef -> Maybe DatumHash

Determine the data script that a transaction output refers to.

updateUtxo :: CardanoTx -> Map TxOutRef TxOut -> Map TxOutRef TxOut

Update a map of unspent transaction outputs and signatures based on the inputs and outputs of a transaction.

txOutPubKey :: TxOut -> Maybe PubKeyHash

pubKeyTxo :: Blockchain -> TxOutRef -> Maybe PubKeyHash

Determine the public key that locks a transaction output, if there is one.

validValuesTx :: Tx -> Bool

Check that all values in a transaction are non-negative.