Ledger.Blockchain

Documentation

newtype OnChainTx

A transaction on the blockchain. Invalid transactions are still put on the chain to be able to collect fees.

Constructors

OnChainTx

Fields

getOnChainTx :: Validated (Tx EmulatorEra)

type Block = [OnChainTx]

A block on the blockchain. This is just a list of transactions following on from the chain so far.

newtype BlockId

Block identifier (usually a hash)

Constructors

BlockId

Fields

getBlockId :: ByteString

type Blockchain = [Block]

A blockchain, which is just a list of blocks, starting with the newest.

newtype Context

Constructors

Context BuiltinData

eitherTx :: (CardanoTx -> r) -> (CardanoTx -> r) -> OnChainTx -> r

unOnChain :: OnChainTx -> CardanoTx

onChainTxIsValid :: OnChainTx -> Bool

consumableInputs :: OnChainTx -> [TxIn]

Outputs consumed from the UTXO set by the OnChainTx

outputsProduced :: OnChainTx -> Map TxIn TxOut

Outputs added to the UTXO set by the OnChainTx