Ledger.Crypto

Documentation

newtype PubKey

Constructors

PubKey

Fields

getPubKey :: LedgerBytes

newtype PrivateKey

A cryptographic private key.

Constructors

PrivateKey

Fields

getPrivateKey :: LedgerBytes

newtype Signature

A message with a cryptographic signature.

Constructors

Signature

Fields

getSignature :: BuiltinByteString

newtype Passphrase

Passphrase newtype to mark intent

Constructors

Passphrase

Fields

unPassphrase :: ByteString

pubKeyHash :: PubKey -> PubKeyHash

Compute the hash of a public key.

signedBy :: ByteArrayAccess a => Signature -> PubKey -> a -> Bool

Check whether the given Signature was signed by the private key corresponding to the given public key.

sign :: ByteArrayAccess a => a -> XPrv -> Passphrase -> Signature

Sign a message using a private key and passphrase.

signTx :: TxId -> XPrv -> Passphrase -> Signature

Sign the hash of a transaction using a private key and passphrase.

generateFromSeed :: ByteString -> Passphrase -> XPrv

Generate a private key from a seed phrase and passphrase

toPublicKey :: XPrv -> PubKey

xPubToPublicKey :: XPub -> PubKey

Signing and generation with no passphrase

sign' :: ByteArrayAccess a => a -> XPrv -> Signature

Sign a message using a private key with no passphrase.

signTx' :: TxId -> XPrv -> Signature

Sign the hash of a transaction using a private key that has no passphrase.

generateFromSeed' :: ByteString -> XPrv

Generate a private key from a seed phrase without a passphrase.