Ledger.Index

Description

An index of unspent transaction outputs, and some functions for validating transactions using the index.

Types for transaction validation based on UTXO index

newtype UtxoIndex

The UTxOs of a blockchain indexed by their references.

Constructors

UtxoIndex

Fields

getIndex :: Map TxOutRef TxOut

insert :: CardanoTx -> UtxoIndex -> UtxoIndex

Update the index for the addition of a transaction.

insertCollateral :: CardanoTx -> UtxoIndex -> UtxoIndex

Update the index for the addition of only the collateral inputs of a failed transaction.

insertBlock :: Block -> UtxoIndex -> UtxoIndex

Update the index for the addition of a block.

initialise :: Blockchain -> UtxoIndex

Create an index of all UTxOs on the chain.

lookup :: MonadError ValidationError m => TxOutRef -> UtxoIndex -> m TxOut

Find an unspent transaction output by the TxOutRef that spends it.

data ValidationError

A reason why a transaction is invalid.

Constructors

TxOutRefNotFound TxOutRef
The transaction output consumed by a transaction input could not be found (either because it was already spent, or because there was no transaction with the given hash on the blockchain).

ScriptFailure ScriptError
For pay-to-script outputs: evaluation of the validator script failed.

CardanoLedgerValidationError Text
An error from Cardano.Ledger validation

_TxOutRefNotFound :: AsValidationError r => Prism' r TxOutRef

_ScriptFailure :: AsValidationError r => Prism' r ScriptError

_CardanoLedgerValidationError :: AsValidationError r => Prism' r Text

type ValidationSuccess = Map RdmrPtr ([Text], ExUnits)

type ValidationErrorInPhase = (ValidationPhase, ValidationError)

data ValidationPhase

Constructors

Phase1

Phase2

minFee :: Tx -> Value

Minimum transaction fee.

maxFee :: Ada

TODO Should be calculated based on the maximum script size permitted on the Cardano blockchain.

adjustTxOut :: PParams (BabbageEra StandardCrypto) -> TxOut -> Either ToCardanoError ([Ada], TxOut)

Adjust a single transaction output so it contains at least the minimum amount of Ada and return the adjustment (if any) and the updated TxOut.

minAdaTxOut :: PParams (BabbageEra StandardCrypto) -> TxOut -> Ada

Exact computation of the mimimum Ada required for a given TxOut. TODO: Should be moved to cardano-api-extended once created

minAdaTxOutEstimated :: Ada

Provide a reasonable estimate of the mimimum of Ada required for a TxOut.

An exact estimate of the the mimimum of Ada in a TxOut is determined by two things: - the PParams, more precisely its coinPerUTxOWord parameter. - the size of the TxOut. In many situations though, we need to determine a plausible value for the minimum of Ada needed for a TxOut without knowing much of the TxOut. This function provides a value big enough to balance UTxOs without a large inlined data (larger than a hash) nor a complex val with a lot of minted values. It's superior to the lowest minimum needed for an UTxO, as the lowest value require no datum. An estimate of the minimum required Ada for each tx output.

maxMinAdaTxOut :: Ada

pubKeyTxIns :: Fold [TxIn] TxIn

Filter to get only the pubkey inputs.

scriptTxIns :: Fold [TxIn] TxIn

Filter to get only the script inputs.

data ExBudget

Constructors

ExBudget

Fields

exBudgetCPU :: ExCPU

exBudgetMemory :: ExMemory

newtype ExCPU

Counts CPU units in picoseconds: maximum value for SatInt is 2^63 ps, or appproximately 106 days.

Constructors

ExCPU CostingInteger

newtype ExMemory

Counts size in machine words.

Constructors

ExMemory CostingInteger

data SatInt