Ledger.Index.Internal

Documentation

newtype UtxoIndex

The UTxOs of a blockchain indexed by their references.

Constructors

UtxoIndex

Fields

getIndex :: Map TxOutRef TxOut

data ValidationError

A reason why a transaction is invalid.

Constructors

TxOutRefNotFound TxOutRef
The transaction output consumed by a transaction input could not be found (either because it was already spent, or because there was no transaction with the given hash on the blockchain).

ScriptFailure ScriptError
For pay-to-script outputs: evaluation of the validator script failed.

CardanoLedgerValidationError Text
An error from Cardano.Ledger validation

class AsValidationError r where

Minimal complete definition

_ValidationError

Methods

_ValidationError :: Prism' r ValidationError

_TxOutRefNotFound :: Prism' r TxOutRef

_ScriptFailure :: Prism' r ScriptError

_CardanoLedgerValidationError :: Prism' r Text

data ValidationPhase

Constructors

Phase1

Phase2

type ValidationErrorInPhase = (ValidationPhase, ValidationError)

type ValidationSuccess = Map RdmrPtr ([Text], ExUnits)