Ledger.Index.Internal

Documentation

newtype OnChainTx

A transaction on the blockchain. Invalid transactions are still put on the chain to be able to collect fees.

Constructors

OnChainTx

Fields

getOnChainTx :: Validated (Tx EmulatorEra)

eitherTx :: (CardanoTx -> r) -> (CardanoTx -> r) -> OnChainTx -> r

unOnChain :: OnChainTx -> CardanoTx

type EmulatorEra = BabbageEra StandardCrypto

type UtxoIndex = UTxO BabbageEra

The UTxOs of a blockchain indexed by their references.

data ValidationError

A reason why a transaction is invalid.

Constructors

TxOutRefNotFound TxIn
The transaction output consumed by a transaction input could not be found (either because it was already spent, or because there was no transaction with the given hash on the blockchain).

ScriptFailure ScriptError
For pay-to-script outputs: evaluation of the validator script failed.

CardanoLedgerValidationError Text
An error from Cardano.Ledger validation

MaxCollateralInputsExceeded
Balancing failed, it needed more than the maximum number of collateral inputs

_MaxCollateralInputsExceeded :: Prism' ValidationError ()

_CardanoLedgerValidationError :: Prism' ValidationError Text

_ScriptFailure :: Prism' ValidationError ScriptError

_TxOutRefNotFound :: Prism' ValidationError TxIn

data ValidationPhase

Constructors

Phase1

Phase2

type ValidationErrorInPhase = (ValidationPhase, ValidationError)

type ValidationSuccess = (RedeemerReport, Validated (Tx EmulatorEra))

type RedeemerReport = Map RdmrPtr ([Text], ExUnits)

data ValidationResult

Constructors

FailPhase1 !CardanoTx !ValidationError
A transaction failed to validate in phase 1.

FailPhase2 !OnChainTx !ValidationError !Value
A transaction failed to validate in phase 2. The Value indicates the amount of collateral stored in the transaction.

Success !OnChainTx !RedeemerReport

_Success :: Prism' ValidationResult (OnChainTx, RedeemerReport)

_FailPhase2 :: Prism' ValidationResult (OnChainTx, ValidationError, Value)

_FailPhase1 :: Prism' ValidationResult (CardanoTx, ValidationError)

data ValidationResultSimple

Constructors

ValidationFailPhase1 !CardanoTx !ValidationError

ValidationFailPhase2 !CardanoTx !ValidationError !Value

ValidationSuccess !CardanoTx

toValidationResultSimple :: ValidationResult -> ValidationResultSimple

cardanoTxFromValidationResult :: ValidationResult -> CardanoTx

toOnChain :: ValidationResult -> Maybe OnChainTx

getEvaluationLogs :: ValidationResult -> [Text]

Get logs from evaluating plutus scripts.

Orphan instances

Semigroup (UTxO era)

Instance details

Methods

(<>) :: UTxO era -> UTxO era -> UTxO era

sconcat :: NonEmpty (UTxO era) -> UTxO era

stimes :: Integral b => b -> UTxO era -> UTxO era

Monoid (UTxO era)

Instance details

Methods

mempty :: UTxO era

mappend :: UTxO era -> UTxO era -> UTxO era

mconcat :: [UTxO era] -> UTxO era