Ledger.Orphans

Orphan instances

Data NetworkId

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> NetworkId -> c NetworkId

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c NetworkId

toConstr :: NetworkId -> Constr

dataTypeOf :: NetworkId -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c NetworkId)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c NetworkId)

gmapT :: (forall b. Data b => b -> b) -> NetworkId -> NetworkId

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> NetworkId -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> NetworkId -> r

gmapQ :: (forall d. Data d => d -> u) -> NetworkId -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> NetworkId -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> NetworkId -> m NetworkId

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> NetworkId -> m NetworkId

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> NetworkId -> m NetworkId

Data NetworkMagic
OpenApi instances for swagger support

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> NetworkMagic -> c NetworkMagic

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c NetworkMagic

toConstr :: NetworkMagic -> Constr

dataTypeOf :: NetworkMagic -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c NetworkMagic)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c NetworkMagic)

gmapT :: (forall b. Data b => b -> b) -> NetworkMagic -> NetworkMagic

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> NetworkMagic -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> NetworkMagic -> r

gmapQ :: (forall d. Data d => d -> u) -> NetworkMagic -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> NetworkMagic -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> NetworkMagic -> m NetworkMagic

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> NetworkMagic -> m NetworkMagic

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> NetworkMagic -> m NetworkMagic

Generic Lovelace

Instance details

Associated Types

type Rep Lovelace :: Type -> Type

Methods

from :: Lovelace -> Rep Lovelace x

to :: Rep Lovelace x -> Lovelace

Generic Quantity

Instance details

Associated Types

type Rep Quantity :: Type -> Type

Methods

from :: Quantity -> Rep Quantity x

to :: Rep Quantity x -> Quantity

Generic PolicyId

Instance details

Associated Types

type Rep PolicyId :: Type -> Type

Methods

from :: PolicyId -> Rep PolicyId x

to :: Rep PolicyId x -> PolicyId

Generic ScriptData

Instance details

Associated Types

type Rep ScriptData :: Type -> Type

Methods

from :: ScriptData -> Rep ScriptData x

to :: Rep ScriptData x -> ScriptData

Generic TxId

Instance details

Associated Types

type Rep TxId :: Type -> Type

Methods

from :: TxId -> Rep TxId x

to :: Rep TxId x -> TxId

Generic TxIx

Instance details

Associated Types

type Rep TxIx :: Type -> Type

Methods

from :: TxIx -> Rep TxIx x

to :: Rep TxIx x -> TxIx

Generic NetworkId

Instance details

Associated Types

type Rep NetworkId :: Type -> Type

Methods

from :: NetworkId -> Rep NetworkId x

to :: Rep NetworkId x -> NetworkId

Hashable POSIXTime

Instance details

Methods

hashWithSalt :: Int -> POSIXTime -> Int

hash :: POSIXTime -> Int

ToJSON ScriptError

Instance details

Methods

toJSON :: ScriptError -> Value

toEncoding :: ScriptError -> Encoding

toJSONList :: [ScriptError] -> Value

toEncodingList :: [ScriptError] -> Encoding

ToJSON POSIXTime
Custom ToJSON instance which allows to simply convert a POSIXTime value to a JSON number.

Instance details

Methods

toJSON :: POSIXTime -> Value

toEncoding :: POSIXTime -> Encoding

toJSONList :: [POSIXTime] -> Value

toEncodingList :: [POSIXTime] -> Encoding

FromJSON ScriptError

Instance details

Methods

parseJSON :: Value -> Parser ScriptError

parseJSONList :: Value -> Parser [ScriptError]

FromJSON POSIXTime
Custom FromJSON instance which allows to parse a JSON number to a POSIXTime value. The parsed JSON value MUST be an Integer or else the parsing fails.

Instance details

Methods

parseJSON :: Value -> Parser POSIXTime

parseJSONList :: Value -> Parser [POSIXTime]

ByteArrayAccess TxId
ByteArrayAccess instance for signing support

Instance details

Methods

length :: TxId -> Int

withByteArray :: TxId -> (Ptr p -> IO a) -> IO a

copyByteArrayToPtr :: TxId -> Ptr p -> IO ()

ToHttpApiData LedgerBytes

Instance details

Methods

toUrlPiece :: LedgerBytes -> Text

toEncodedUrlPiece :: LedgerBytes -> Builder

toHeader :: LedgerBytes -> ByteString

toQueryParam :: LedgerBytes -> Text

ToHttpApiData PrivateKey

Instance details

Methods

toUrlPiece :: PrivateKey -> Text

toEncodedUrlPiece :: PrivateKey -> Builder

toHeader :: PrivateKey -> ByteString

toQueryParam :: PrivateKey -> Text

FromHttpApiData LedgerBytes

Instance details

Methods

parseUrlPiece :: Text -> Either Text LedgerBytes

parseHeader :: ByteString -> Either Text LedgerBytes

parseQueryParam :: Text -> Either Text LedgerBytes

FromHttpApiData PrivateKey

Instance details

Methods

parseUrlPiece :: Text -> Either Text PrivateKey

parseHeader :: ByteString -> Either Text PrivateKey

parseQueryParam :: Text -> Either Text PrivateKey

ToSchema Value

Instance details

Methods

declareNamedSchema :: Proxy Value -> Declare (Definitions Schema) NamedSchema

ToSchema XPrv

Instance details

Methods

declareNamedSchema :: Proxy XPrv -> Declare (Definitions Schema) NamedSchema

ToSchema Lovelace

Instance details

Methods

declareNamedSchema :: Proxy Lovelace -> Declare (Definitions Schema) NamedSchema

ToSchema Quantity

Instance details

Methods

declareNamedSchema :: Proxy Quantity -> Declare (Definitions Schema) NamedSchema

ToSchema PolicyId

Instance details

Methods

declareNamedSchema :: Proxy PolicyId -> Declare (Definitions Schema) NamedSchema

ToSchema AssetName

Instance details

Methods

declareNamedSchema :: Proxy AssetName -> Declare (Definitions Schema) NamedSchema

ToSchema ScriptHash

Instance details

Methods

declareNamedSchema :: Proxy ScriptHash -> Declare (Definitions Schema) NamedSchema

ToSchema ScriptData

Instance details

Methods

declareNamedSchema :: Proxy ScriptData -> Declare (Definitions Schema) NamedSchema

ToSchema TxId

Instance details

Methods

declareNamedSchema :: Proxy TxId -> Declare (Definitions Schema) NamedSchema

ToSchema TxIx

Instance details

Methods

declareNamedSchema :: Proxy TxIx -> Declare (Definitions Schema) NamedSchema

ToSchema XPub

Instance details

Methods

declareNamedSchema :: Proxy XPub -> Declare (Definitions Schema) NamedSchema

ToSchema Language

Instance details

Methods

declareNamedSchema :: Proxy Language -> Declare (Definitions Schema) NamedSchema

ToSchema LogLevel

Instance details

Methods

declareNamedSchema :: Proxy LogLevel -> Declare (Definitions Schema) NamedSchema

ToSchema TxOut

Instance details

Methods

declareNamedSchema :: Proxy TxOut -> Declare (Definitions Schema) NamedSchema

ToSchema TxId

Instance details

Methods

declareNamedSchema :: Proxy TxId -> Declare (Definitions Schema) NamedSchema

ToSchema ScriptTag

Instance details

Methods

declareNamedSchema :: Proxy ScriptTag -> Declare (Definitions Schema) NamedSchema

ToSchema RedeemerPtr

Instance details

Methods

declareNamedSchema :: Proxy RedeemerPtr -> Declare (Definitions Schema) NamedSchema

ToSchema TxOutRef

Instance details

Methods

declareNamedSchema :: Proxy TxOutRef -> Declare (Definitions Schema) NamedSchema

ToSchema TxOut

Instance details

Methods

declareNamedSchema :: Proxy TxOut -> Declare (Definitions Schema) NamedSchema

ToSchema DCert

Instance details

Methods

declareNamedSchema :: Proxy DCert -> Declare (Definitions Schema) NamedSchema

ToSchema CurrencySymbol

Instance details

Methods

declareNamedSchema :: Proxy CurrencySymbol -> Declare (Definitions Schema) NamedSchema

ToSchema TokenName

Instance details

Methods

declareNamedSchema :: Proxy TokenName -> Declare (Definitions Schema) NamedSchema

ToSchema AssetClass

Instance details

Methods

declareNamedSchema :: Proxy AssetClass -> Declare (Definitions Schema) NamedSchema

ToSchema Value

Instance details

Methods

declareNamedSchema :: Proxy Value -> Declare (Definitions Schema) NamedSchema

ToSchema Script

Instance details

Methods

declareNamedSchema :: Proxy Script -> Declare (Definitions Schema) NamedSchema

ToSchema Validator

Instance details

Methods

declareNamedSchema :: Proxy Validator -> Declare (Definitions Schema) NamedSchema

ToSchema Redeemer

Instance details

Methods

declareNamedSchema :: Proxy Redeemer -> Declare (Definitions Schema) NamedSchema

ToSchema MintingPolicy

Instance details

Methods

declareNamedSchema :: Proxy MintingPolicy -> Declare (Definitions Schema) NamedSchema

ToSchema StakeValidator

Instance details

Methods

declareNamedSchema :: Proxy StakeValidator -> Declare (Definitions Schema) NamedSchema

ToSchema ScriptHash

Instance details

Methods

declareNamedSchema :: Proxy ScriptHash -> Declare (Definitions Schema) NamedSchema

ToSchema RedeemerHash

Instance details

Methods

declareNamedSchema :: Proxy RedeemerHash -> Declare (Definitions Schema) NamedSchema

ToSchema MintingPolicyHash

Instance details

Methods

declareNamedSchema :: Proxy MintingPolicyHash -> Declare (Definitions Schema) NamedSchema

ToSchema LedgerBytes

Instance details

Methods

declareNamedSchema :: Proxy LedgerBytes -> Declare (Definitions Schema) NamedSchema

ToSchema POSIXTime

Instance details

Methods

declareNamedSchema :: Proxy POSIXTime -> Declare (Definitions Schema) NamedSchema

ToSchema DiffMilliSeconds

Instance details

Methods

declareNamedSchema :: Proxy DiffMilliSeconds -> Declare (Definitions Schema) NamedSchema

ToSchema Ada

Instance details

Methods

declareNamedSchema :: Proxy Ada -> Declare (Definitions Schema) NamedSchema

ToSchema Slot

Instance details

Methods

declareNamedSchema :: Proxy Slot -> Declare (Definitions Schema) NamedSchema

ToSchema Signature

Instance details

Methods

declareNamedSchema :: Proxy Signature -> Declare (Definitions Schema) NamedSchema

ToSchema PubKey

Instance details

Methods

declareNamedSchema :: Proxy PubKey -> Declare (Definitions Schema) NamedSchema

Serialise POSIXTime

Instance details

Methods

encode :: POSIXTime -> Encoding

decode :: Decoder s POSIXTime

encodeList :: [POSIXTime] -> Encoding

decodeList :: Decoder s [POSIXTime]

ToSchema (AddressInEra BabbageEra)

Instance details

Methods

declareNamedSchema :: Proxy (AddressInEra BabbageEra) -> Declare (Definitions Schema) NamedSchema

ToSchema (Hash ScriptData)

Instance details

Methods

declareNamedSchema :: Proxy (Hash ScriptData) -> Declare (Definitions Schema) NamedSchema

ToSchema (Digest Blake2b_160)

Instance details

Methods

declareNamedSchema :: Proxy (Digest Blake2b_160) -> Declare (Definitions Schema) NamedSchema

ToSchema (LogMessage Value)

Instance details

Methods

declareNamedSchema :: Proxy (LogMessage Value) -> Declare (Definitions Schema) NamedSchema

ToSchema ann => ToSchema (Kind ann)

Instance details

Methods

declareNamedSchema :: Proxy (Kind ann) -> Declare (Definitions Schema) NamedSchema

ToSchema ann => ToSchema (Version ann)

Instance details

Methods

declareNamedSchema :: Proxy (Version ann) -> Declare (Definitions Schema) NamedSchema

ToSchema a => ToSchema (Interval a)

Instance details

Methods

declareNamedSchema :: Proxy (Interval a) -> Declare (Definitions Schema) NamedSchema

ToSchema a => ToSchema (Extended a)

Instance details

Methods

declareNamedSchema :: Proxy (Extended a) -> Declare (Definitions Schema) NamedSchema

ToSchema a => ToSchema (UpperBound a)

Instance details

Methods

declareNamedSchema :: Proxy (UpperBound a) -> Declare (Definitions Schema) NamedSchema

ToSchema a => ToSchema (LowerBound a)

Instance details

Methods

declareNamedSchema :: Proxy (LowerBound a) -> Declare (Definitions Schema) NamedSchema

ToSchema script => ToSchema (Versioned script)

Instance details

Methods

declareNamedSchema :: Proxy (Versioned script) -> Declare (Definitions Schema) NamedSchema

Serialise (AddressInEra BabbageEra)

Instance details

Methods

encode :: AddressInEra BabbageEra -> Encoding

decode :: Decoder s (AddressInEra BabbageEra)

encodeList :: [AddressInEra BabbageEra] -> Encoding

decodeList :: Decoder s [AddressInEra BabbageEra]

ToSchema (Hash Blake2b_256 EraIndependentTxBody)

Instance details

Methods

declareNamedSchema :: Proxy (Hash Blake2b_256 EraIndependentTxBody) -> Declare (Definitions Schema) NamedSchema

ToSchema (SafeHash StandardCrypto EraIndependentData)

Instance details

Methods

declareNamedSchema :: Proxy (SafeHash StandardCrypto EraIndependentData) -> Declare (Definitions Schema) NamedSchema

(ToSchema k, ToSchema v) => ToSchema (Map k v)

Instance details

Methods

declareNamedSchema :: Proxy (Map k v) -> Declare (Definitions Schema) NamedSchema

(ToSchema tyname, ToSchema name, ToSchema (uni ann), ToSchema fun, ToSchema ann, ToSchema (Type tyname uni ann), ToSchema (Some (ValueOf uni)), Typeable uni) => ToSchema (Term tyname name uni fun ann)

Instance details

Methods

declareNamedSchema :: Proxy (Term tyname name uni fun ann) -> Declare (Definitions Schema) NamedSchema