Ledger.Tx

Documentation

data TxOutRef

Constructors

TxOutRef

Fields

txOutRefId :: TxId

txOutRefIdx :: Integer

inScripts :: TxIn -> Maybe (Validator, Redeemer, Datum)

isPayToScriptOut :: TxOut -> Bool

isPubKeyOut :: TxOut -> Bool

pubKeyHashTxOut :: Value -> PubKeyHash -> TxOut

pubKeyTxIns :: Fold (Set TxIn) TxIn

scriptTxIns :: Fold (Set TxIn) TxIn

data RedeemerPtr

Constructors

RedeemerPtr ScriptTag Integer

type Redeemers = Map RedeemerPtr Redeemer

data ScriptTag

Constructors

Spend

Mint

Cert

Reward

data Language

Constructors

PlutusV1

PlutusV2

data Versioned script

Constructors

Versioned

Fields

unversioned :: script

version :: Language

type ReferenceScript = ReferenceScript BabbageEra

type MintingWitnessesMap = Map MintingPolicyHash (Redeemer, Maybe (Versioned TxOutRef))

type ScriptsMap = Map ScriptHash (Versioned Script)

newtype TxOut

Constructors

TxOut

Fields

getTxOut :: TxOut CtxTx BabbageEra

data Certificate

Constructors

Certificate

Fields

certificateDcert :: DCert

certificateRedeemer :: Maybe Redeemer

redeemer for script credential

data Withdrawal

Stake withdrawal, if applicable the script should be included in txScripts.

Constructors

Withdrawal

Fields

withdrawalCredential :: Credential

staking credential

withdrawalAmount :: Integer

amount of withdrawal in Lovelace, must withdraw all eligible amount

withdrawalRedeemer :: Maybe Redeemer

redeemer for script credential

cardanoTxOutValue :: TxOut ctx era -> Value

txOutValue :: TxOut -> Value

outValue :: Lens TxOut TxOut Value (TxOutValue BabbageEra)

outValue' :: Lens' TxOut (TxOutValue BabbageEra)

toSizedTxOut :: TxOut -> Sized (TxOut StandardBabbage)

toCtxUTxOTxOut :: TxOut -> TxOut CtxUTxO BabbageEra

txOutDatumHash :: TxOut -> Maybe DatumHash

Get a hash from the stored TxOutDatum (either directly or by hashing the inlined datum)

txOutDatum :: forall d. FromData d => TxOut -> Maybe d

cardanoTxOutDatumHash :: TxOutDatum CtxUTxO BabbageEra -> Maybe (Hash ScriptData)

txOutPubKey :: TxOut -> Maybe PubKeyHash

txOutAddress :: TxOut -> CardanoAddress

outAddress :: Lens' TxOut (AddressInEra BabbageEra)

outDatumHash :: Lens TxOut TxOut (Maybe DatumHash) (TxOutDatum CtxTx BabbageEra)

txOutReferenceScript :: TxOut -> ReferenceScript

outReferenceScript :: Lens' TxOut ReferenceScript

lookupScript :: ScriptsMap -> ScriptHash -> Maybe (Versioned Script)

lookupValidator :: ScriptsMap -> ValidatorHash -> Maybe (Versioned Validator)

lookupMintingPolicy :: ScriptsMap -> MintingPolicyHash -> Maybe (Versioned MintingPolicy)

lookupStakeValidator :: ScriptsMap -> StakeValidatorHash -> Maybe (Versioned StakeValidator)

emptyTxBodyContent :: TxBodyContent BuildTx BabbageEra

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

data TxIn

Constructors

TxIn TxId TxIx

newtype TxIx

Constructors

TxIx Word

DecoratedTxOut

data DecoratedTxOut

Offchain view of a transaction output.

Constructors

PublicKeyDecoratedTxOut

Fields

_decoratedTxOutPubKeyHash :: PubKeyHash

The pubKey hash that protects the transaction address

_decoratedTxOutStakingCredential :: Maybe StakingCredential

The staking credential of the transaction address, if any

_decoratedTxOutValue :: Value

Value of the transaction output.

_decoratedTxOutPubKeyDatum :: Maybe (DatumHash, DatumFromQuery)

Optional datum (inline datum or datum in transaction body) attached to the transaction output.

_decoratedTxOutReferenceScript :: Maybe (Versioned Script)

Value of the transaction output.

ScriptDecoratedTxOut

Fields

_decoratedTxOutValidatorHash :: ValidatorHash

The hash of the script that protects the transaction address

_decoratedTxOutStakingCredential :: Maybe StakingCredential

The staking credential of the transaction address, if any

_decoratedTxOutValue :: Value

Value of the transaction output.

_decoratedTxOutScriptDatum :: (DatumHash, DatumFromQuery)

Datum attached to the transaction output, either in full (inline datum or datum in transaction body) or as a hash reference. A transaction output protected by a Plutus script is guardateed to have an associated datum.

_decoratedTxOutReferenceScript :: Maybe (Versioned Script)

Value of the transaction output.

_decoratedTxOutValidator :: Maybe (Versioned Validator)

Full version of the validator protecting the transaction output

toTxOut :: NetworkId -> DecoratedTxOut -> Either ToCardanoError TxOut

toTxInfoTxOut :: DecoratedTxOut -> TxOut

Converts a transaction output from the chain index to the plutus-ledger-api transaction output.

Note that DecoratedTxOut supports features such inline datums and reference scripts which are not supported by V1 TxOut. Converting from DecoratedTxOut to TxOut and back is therefore lossy.

toDecoratedTxOut :: TxOut -> Maybe DecoratedTxOut

Lenses and Prisms

decoratedTxOutPubKeyHash :: Traversal' DecoratedTxOut PubKeyHash

decoratedTxOutAddress :: Getter DecoratedTxOut Address

decoratedTxOutDatum :: Traversal' DecoratedTxOut (DatumHash, DatumFromQuery)

decoratedTxOutValue :: Lens' DecoratedTxOut Value

decoratedTxOutPubKeyDatum :: Traversal' DecoratedTxOut (Maybe (DatumHash, DatumFromQuery))

decoratedTxOutScriptDatum :: Traversal' DecoratedTxOut (DatumHash, DatumFromQuery)

decoratedTxOutStakingCredential :: Lens' DecoratedTxOut (Maybe StakingCredential)

decoratedTxOutReferenceScript :: Lens' DecoratedTxOut (Maybe (Versioned Script))

decoratedTxOutValidatorHash :: Traversal' DecoratedTxOut ValidatorHash

decoratedTxOutValidator :: Traversal' DecoratedTxOut (Maybe (Versioned Validator))

_PublicKeyDecoratedTxOut :: Prism' DecoratedTxOut (PubKeyHash, Maybe StakingCredential, Value, Maybe (DatumHash, DatumFromQuery), Maybe (Versioned Script))

_ScriptDecoratedTxOut :: Prism' DecoratedTxOut (ValidatorHash, Maybe StakingCredential, Value, (DatumHash, DatumFromQuery), Maybe (Versioned Script), Maybe (Versioned Validator))

_decoratedTxOutAddress :: DecoratedTxOut -> Address

smart Constructors

mkDecoratedTxOut :: CardanoAddress -> Value -> Maybe (DatumHash, DatumFromQuery) -> Maybe (Versioned Script) -> Maybe DecoratedTxOut

mkPubkeyDecoratedTxOut :: CardanoAddress -> Value -> Maybe (DatumHash, DatumFromQuery) -> Maybe (Versioned Script) -> Maybe DecoratedTxOut

mkScriptDecoratedTxOut :: CardanoAddress -> Value -> (DatumHash, DatumFromQuery) -> Maybe (Versioned Script) -> Maybe (Versioned Validator) -> Maybe DecoratedTxOut

DatumFromQuery

data DatumFromQuery

A datum in a transaction output that comes from a chain index query.

Constructors

DatumUnknown

DatumInline Datum

DatumInBody Datum

datumInDatumFromQuery :: Traversal' DatumFromQuery Datum

Transactions

getCardanoTxId :: CardanoTx -> TxId

getCardanoTxInputs :: CardanoTx -> [TxIn]

getCardanoTxCollateralInputs :: CardanoTx -> [TxIn]

getCardanoTxOutRefs :: CardanoTx -> [(TxOut, TxIn)]

getCardanoTxOutputs :: CardanoTx -> [TxOut]

getCardanoTxRedeemers :: CardanoTx -> Redeemers

getCardanoTxSpentOutputs :: CardanoTx -> Set TxIn

getCardanoTxProducedOutputs :: CardanoTx -> Map TxIn TxOut

getCardanoTxReturnCollateral :: CardanoTx -> Maybe TxOut

getCardanoTxProducedReturnCollateral :: CardanoTx -> Map TxIn TxOut

getCardanoTxTotalCollateral :: CardanoTx -> Maybe Lovelace

getCardanoTxFee :: CardanoTx -> Lovelace

getCardanoTxMint :: CardanoTx -> Value

getCardanoTxValidityRange :: CardanoTx -> SlotRange

getCardanoTxData :: CardanoTx -> Map DatumHash Datum

data CardanoTx where

Cardano tx from any era.

Constructors

CardanoTx :: IsCardanoEra era => Tx era -> EraInMode era CardanoMode -> CardanoTx

Bundled Patterns

pattern CardanoEmulatorEraTx :: Tx BabbageEra -> CardanoTx

data ToCardanoError

Constructors

TxBodyError String
A C.TxBodyError converted to String

DeserialisationError

InvalidValidityRange

ValueNotPureAda

OutputHasZeroAda

StakingPointersNotSupported

SimpleScriptsNotSupportedToCardano

MissingInputValidator

MissingDatum

MissingMintingPolicy

ScriptPurposeNotSupported ScriptTag

MissingMintingPolicyRedeemer

MissingStakeValidator

UnsupportedPlutusVersion Language

Tag String ToCardanoError

addCardanoTxSignature :: PrivateKey -> CardanoTx -> CardanoTx

TxBodyContent functions

getTxBodyContentInputs :: TxBodyContent ctx era -> [TxIn]

getTxBodyContentCollateralInputs :: TxBodyContent ctx era -> [TxIn]

getTxBodyContentReturnCollateral :: TxBodyContent ctx BabbageEra -> Maybe TxOut

getTxBodyContentMint :: TxBodyContent ctx era -> Value

txBodyContentIns :: Lens' (TxBodyContent BuildTx BabbageEra) [(TxIn, BuildTxWith BuildTx (Witness WitCtxTxIn BabbageEra))]

txBodyContentCollateralIns :: Lens' (TxBodyContent BuildTx BabbageEra) [TxIn]

txBodyContentOuts :: Lens' (TxBodyContent ctx BabbageEra) [TxOut]

Utility

decoratedTxOutPlutusValue :: DecoratedTxOut -> Value

fromDecoratedIndex :: NetworkId -> Map TxOutRef DecoratedTxOut -> Either ToCardanoError UtxoIndex

Orphan instances

Pretty CardanoTx

Instance details

Methods

pretty :: CardanoTx -> Doc ann

prettyList :: [CardanoTx] -> Doc ann

Pretty CardanoBuildTx

Instance details

Methods

pretty :: CardanoBuildTx -> Doc ann

prettyList :: [CardanoBuildTx] -> Doc ann