Ledger.Tx.CardanoAPI.Internal

Description

Interface to the transaction types from 'cardano-api'

Documentation

newtype CardanoBuildTx

Constructors

CardanoBuildTx

Fields

getCardanoBuildTx :: TxBodyContent BuildTx BabbageEra

data SomeCardanoApiTx where

Cardano tx from any era.

Constructors

SomeTx :: IsCardanoEra era => Tx era -> EraInMode era CardanoMode -> SomeCardanoApiTx

txOutRefs :: SomeCardanoApiTx -> [(TxOut, TxOutRef)]

unspentOutputsTx :: SomeCardanoApiTx -> Map TxOutRef TxOut

fromCardanoTxId :: TxId -> TxId

fromCardanoTxIn :: TxIn -> TxOutRef

fromCardanoTxOutToPV1TxInfoTxOut :: TxOut CtxTx era -> TxOut

fromCardanoTxOutToPV2TxInfoTxOut :: TxOut CtxTx era -> TxOut

fromCardanoTxOutDatumHash :: TxOutDatum CtxTx era -> Maybe DatumHash

fromCardanoTxOutDatum :: TxOutDatum CtxTx era -> OutputDatum

fromCardanoTxOutValue :: TxOutValue era -> Value

fromCardanoAddressInEra :: AddressInEra era -> Address

Deprecated: we now use Cardano address internally, if you need a plutus address use toPlutusAddress

fromCardanoAddress :: Address addrtype -> Address

Deprecated: Shouldn't be used as we use Cardano address internally now

fromCardanoAssetId :: AssetId -> AssetClass

fromCardanoAssetName :: AssetName -> TokenName

fromCardanoMintValue :: TxMintValue build era -> Value

fromCardanoValue :: Value -> Value

fromCardanoPolicyId :: PolicyId -> MintingPolicyHash

fromCardanoFee :: TxFee era -> Value

fromCardanoValidityRange :: (TxValidityLowerBound era, TxValidityUpperBound era) -> SlotRange

fromCardanoScriptInEra :: ScriptInEra era -> Maybe (Versioned Script)

fromCardanoPaymentKeyHash :: Hash PaymentKey -> PubKeyHash

Deprecated: Shouldn't be used as we use Cardano address internally now

fromCardanoScriptData :: ScriptData -> BuiltinData

fromCardanoPlutusScript :: HasTypeProxy lang => PlutusScript lang -> Script

fromCardanoScriptInAnyLang :: ScriptInAnyLang -> Maybe (Versioned Script)

fromCardanoLovelace :: Lovelace -> Value

fromTxScriptValidity :: TxScriptValidity era -> Bool

Given a 'C.TxScriptValidity era', if the era supports scripts, return a True or False depending on script validity. If the era does not support scripts, always return True.

toTxScriptValidity :: ShelleyBasedEra era -> Bool -> TxScriptValidity era

scriptDataFromCardanoTxBody :: TxBody era -> (Map DatumHash Datum, Redeemers)

Given a 'C.TxBody from a 'C.Tx era', return the datums and redeemers along with their hashes.

plutusScriptsFromTxBody :: TxBody era -> Map ScriptHash (Versioned Script)

Extract plutus scripts from a Cardano API tx body.

Note that Plutus scripts are only supported in Alonzo era and onwards.

makeTransactionBody :: Maybe (PParams (BabbageEra StandardCrypto)) -> Map RdmrPtr ExUnits -> CardanoBuildTx -> Either ToCardanoError (TxBody BabbageEra)

toCardanoTxIn :: TxOutRef -> Either ToCardanoError TxIn

toCardanoTxOut :: NetworkId -> TxOut -> Either ToCardanoError (TxOut CtxTx BabbageEra)

toCardanoTxOutDatum :: OutputDatum -> Either ToCardanoError (TxOutDatum CtxTx BabbageEra)

toCardanoTxOutDatumHash :: DatumHash -> Either ToCardanoError (TxOutDatum ctx BabbageEra)

toCardanoTxOutDatumHashFromDatum :: Datum -> TxOutDatum ctx BabbageEra

toCardanoTxOutDatumInline :: Datum -> TxOutDatum CtxTx BabbageEra

toCardanoTxOutDatumInTx :: Datum -> TxOutDatum CtxTx BabbageEra

toCardanoTxOutNoDatum :: TxOutDatum CtxTx BabbageEra

toCardanoTxOutValue :: Value -> Either ToCardanoError (TxOutValue BabbageEra)

toCardanoAddressInEra :: NetworkId -> Address -> Either ToCardanoError (AddressInEra BabbageEra)

toCardanoAssetId :: AssetClass -> Either ToCardanoError AssetId

toCardanoAssetName :: TokenName -> Either ToCardanoError AssetName

toCardanoPolicyId :: MintingPolicyHash -> Either ToCardanoError PolicyId

toCardanoValue :: Value -> Either ToCardanoError Value

toCardanoLovelace :: Value -> Either ToCardanoError Lovelace

toCardanoFee :: Value -> Either ToCardanoError (TxFee BabbageEra)

adaToCardanoValue :: Ada -> Value

toCardanoValidityRange :: SlotRange -> Either ToCardanoError (TxValidityLowerBound BabbageEra, TxValidityUpperBound BabbageEra)

toCardanoScriptInEra :: Versioned Script -> Either ToCardanoError (ScriptInEra BabbageEra)

toCardanoPaymentKeyHash :: PaymentPubKeyHash -> Either ToCardanoError (Hash PaymentKey)

toCardanoScriptData :: BuiltinData -> ScriptData

toCardanoScriptDataHash :: DatumHash -> Either ToCardanoError (Hash ScriptData)

toCardanoScriptHash :: ValidatorHash -> Either ToCardanoError ScriptHash

toCardanoStakeKeyHash :: PubKeyHash -> Either ToCardanoError (Hash StakeKey)

toCardanoPlutusScript :: SerialiseAsRawBytes plutusScript => AsType plutusScript -> Script -> Either ToCardanoError plutusScript

toCardanoScriptInAnyLang :: Versioned Script -> Either ToCardanoError ScriptInAnyLang

toCardanoReferenceScript :: Maybe (Versioned Script) -> Either ToCardanoError (ReferenceScript BabbageEra)

toCardanoTxId :: TxId -> Either ToCardanoError TxId

data ToCardanoError

Constructors

TxBodyError String
A C.TxBodyError converted to String

DeserialisationError

InvalidValidityRange

ValueNotPureAda

OutputHasZeroAda

StakingPointersNotSupported

SimpleScriptsNotSupportedToCardano

MissingInputValidator

MissingDatum

MissingMintingPolicy

ScriptPurposeNotSupported ScriptTag

MissingMintingPolicyRedeemer

MissingStakeValidator

UnsupportedPlutusVersion Language

Tag String ToCardanoError

data FromCardanoError

Constructors

SimpleScriptsNotSupported

deserialiseFromRawBytes :: SerialiseAsRawBytes t => AsType t -> ByteString -> Either ToCardanoError t

zeroExecutionUnits :: ExecutionUnits

tag :: String -> Either ToCardanoError t -> Either ToCardanoError t

withIsCardanoEra :: EraInMode era CardanoMode -> (IsCardanoEra era => r) -> r

Run code that needs an IsCardanoEra constraint while you only have an EraInMode value.

Orphan instances

Typeable era => ToSchema (Tx era)

Instance details

Methods

declareNamedSchema :: Proxy (Tx era) -> Declare (Definitions Schema) NamedSchema

(Typeable era, Typeable mode) => ToSchema (EraInMode era mode)

Instance details

Methods

declareNamedSchema :: Proxy (EraInMode era mode) -> Declare (Definitions Schema) NamedSchema