Ledger.Tx.Internal

Documentation

type ReferenceScript = ReferenceScript BabbageEra

data TxOutTx

A TxOut along with the Tx it comes from, which may have additional information e.g. the full data script that goes with the TxOut.

Constructors

TxOutTx

Fields

txOutTxTx :: Tx

txOutTxOut :: TxOut

data TxStripped

A babbage era transaction without witnesses for its inputs.

Constructors

TxStripped

Fields

txStrippedInputs :: [TxOutRef]

The inputs to this transaction, as transaction output references only.

txStrippedReferenceInputs :: [TxOutRef]

The reference inputs to this transaction, as transaction output references only.

txStrippedOutputs :: [TxOut]

The outputs of this transation.

txStrippedMint :: !Value

The Value minted by this transaction.

txStrippedFee :: !Value

The fee for this transaction.

data Tx

A Babbage-era transaction, including witnesses for its inputs.

Constructors

Tx

Fields

txInputs :: [TxInput]

The inputs to this transaction.

txReferenceInputs :: [TxInput]

The reference inputs to this transaction.

txCollateralInputs :: [TxInput]

The collateral inputs to cover the fees in case validation of the transaction fails.

txOutputs :: [TxOut]

The outputs of this transaction, ordered so they can be referenced by index.

txReturnCollateral :: Maybe TxOut

The output of the remaining collateral after covering fees in case validation of the transaction fails.

txTotalCollateral :: Maybe Value

The total collateral to be paid in case validation of the transaction fails.

txMint :: !Value

The Value minted by this transaction.

txFee :: !Value

The fee for this transaction.

txValidRange :: !SlotRange

The SlotRange during which this transaction may be validated.

txMintingWitnesses :: MintingWitnessesMap

The witnesses that must be present to check minting conditions matched with their redeemers.

txWithdrawals :: [Withdrawal]

Withdrawals, contains redeemers.

txCertificates :: [Certificate]

Certificates, contains redeemers.

txSignatures :: Map PubKey Signature

Signatures of this transaction.

txScripts :: ScriptsMap

Scripts for all script credentials mentioned in this tx.

txData :: Map DatumHash Datum

Datum objects recorded on this transaction.

txMetadata :: Maybe BuiltinByteString

Metadata

type MintingWitnessesMap = Map MintingPolicyHash (Redeemer, Maybe (Versioned TxOutRef))

type ScriptsMap = Map ScriptHash (Versioned Script)

newtype TxOut

Validator, redeemer, and data scripts of a transaction input that spends a "pay to script" output. inScripts :: Tx -> TxInput -> Maybe (LedgerPlutusVersion, Validator, Redeemer, Datum) inScripts tx i@TxInput{txInputType=TxConsumeScriptAddress pv _ _ _} = case txInType $ fillTxInputWitnesses tx i of Just (ConsumeScriptAddress v r d) -> Just (pv, v, r, d) _ -> Nothing inScripts _ _ = Nothing

Constructors

TxOut

Fields

getTxOut :: TxOut CtxTx BabbageEra

data Certificate

Constructors

Certificate

Fields

certificateDcert :: DCert

certificateRedeemer :: Maybe Redeemer

redeemer for script credential

data Withdrawal

Stake withdrawal, if applicable the script should be included in txScripts.

Constructors

Withdrawal

Fields

withdrawalCredential :: Credential

staking credential

withdrawalAmount :: Integer

amount of withdrawal in Lovelace, must withdraw all eligible amount

withdrawalRedeemer :: Maybe Redeemer

redeemer for script credential

data TxInput

A transaction input, consisting of a transaction output reference and an input type. Differs with TxIn by: TxIn *maybe* contains *full* data witnesses, TxInput always contains redeemer witness, but datum/validator hashes.

Constructors

TxInput

Fields

txInputRef :: !TxOutRef

txInputType :: !TxInputType

data TxInputType

The type of a transaction input with hashes.

Constructors

TxScriptAddress !Redeemer !(Either ValidatorHash (Versioned TxOutRef)) !(Maybe DatumHash)
A transaction input that consumes (with a validator hash) or references (with a txOutRef) a script address with the given the redeemer and datum hash.

TxConsumePublicKeyAddress
A transaction input that consumes a public key address.

TxConsumeSimpleScriptAddress
Consume a simple script

data TxIn

A transaction input, consisting of a transaction output reference and an input type.

Constructors

TxIn

Fields

txInRef :: !TxOutRef

txInType :: Maybe TxInType

data TxInType

The type of a transaction input.

Constructors

ScriptAddress !(Either (Versioned Validator) (Versioned TxOutRef)) !Redeemer !(Maybe Datum)
A transaction input that consumes (with a validator) or references (with a txOutRef) a script address with the given the redeemer and datum. Datum is optional if the input refers to a script output which contains an inline datum

ConsumePublicKeyAddress
A transaction input that consumes a public key address.

ConsumeSimpleScriptAddress
Consume a simple script

pubKeyTxIn :: TxOutRef -> TxIn

A transaction input that spends a "pay to public key" output, given the witness.

scriptTxIn :: TxOutRef -> Versioned Validator -> Redeemer -> Maybe Datum -> TxIn

A transaction input that spends a "pay to script" output, given witnesses. Datum is optional if the input refers to a script output which contains an inline datum

inputRef :: Lens' TxInput TxOutRef

The TxOutRef spent by a transaction input.

inputType :: Lens' TxInput TxInputType

The type of a transaction input.

inScripts :: TxIn -> Maybe (Versioned Validator, Redeemer, Maybe Datum)

Validator, redeemer, and data scripts of a transaction input that spends a "pay to script" output.

inRef :: Lens' TxInput TxOutRef

The TxOutRef spent by a transaction input.

inType :: Lens' TxInput TxInputType

The type of a transaction input.

pubKeyTxInputs :: Fold [TxInput] TxInput

Filter to get only the pubkey inputs.

scriptTxInputs :: Fold [TxInput] TxInput

Filter to get only the scripts that consume or reference a script address

referenceScriptTxInputs :: Fold [TxInput] TxInput

Filter to get only the scripts that reference a script address

toSizedTxOut :: TxOut -> Sized (TxOut StandardBabbage)

inputs :: Lens' Tx [TxInput]

The inputs of a transaction.

referenceInputs :: Lens' Tx [TxInput]

The reference inputs of a transaction.

collateralInputs :: Lens' Tx [TxInput]

The collateral inputs of a transaction for paying fees when validating the transaction fails.

outputs :: Lens' Tx [TxOut]

The outputs of a transaction.

returnCollateral :: Lens' Tx (Maybe TxOut)

totalCollateral :: Lens' Tx (Maybe Value)

validRange :: Lens' Tx SlotRange

The validity range of a transaction.

signatures :: Lens' Tx (Map PubKey Signature)

fee :: Lens' Tx Value

mint :: Lens' Tx Value

mintScripts :: Lens' Tx MintingWitnessesMap

scriptWitnesses :: Lens' Tx ScriptsMap

datumWitnesses :: Lens' Tx (Map DatumHash Datum)

metadata :: Lens' Tx (Maybe BuiltinByteString)

The inputs of a transaction.

lookupSignature :: PubKey -> Tx -> Maybe Signature

lookupDatum :: Tx -> DatumHash -> Maybe Datum

validValuesTx :: Tx -> Bool

Check that all values in a transaction are non-negative.

txOutValue :: TxOut -> Value

outValue :: Lens TxOut TxOut Value (TxOutValue BabbageEra)

outValue' :: Lens' TxOut (TxOutValue BabbageEra)

strip :: Tx -> TxStripped

txOutTxDatum :: TxOutTx -> Maybe Datum

txOutDatumHash :: TxOut -> Maybe DatumHash

Get a hash from the stored TxOutDatum (either dirctly or by hashing the inlined datum)

txOutPubKey :: TxOut -> Maybe PubKeyHash

txOutAddress :: TxOut -> CardanoAddress

outAddress :: Lens' TxOut (AddressInEra BabbageEra)

outDatumHash :: Lens TxOut TxOut (Maybe DatumHash) (TxOutDatum CtxTx BabbageEra)

txOutReferenceScript :: TxOut -> ReferenceScript

outReferenceScript :: Lens' TxOut ReferenceScript

lookupScript :: ScriptsMap -> ScriptHash -> Maybe (Versioned Script)

lookupValidator :: ScriptsMap -> ValidatorHash -> Maybe (Versioned Validator)

spentOutputs :: Tx -> [TxOutRef]

The transaction output references consumed by a transaction.

referencedOutputs :: Tx -> [TxOutRef]

The transaction output references referenced by a transaction.

lookupMintingPolicy :: ScriptsMap -> MintingPolicyHash -> Maybe (Versioned MintingPolicy)

lookupStakeValidator :: ScriptsMap -> StakeValidatorHash -> Maybe (Versioned StakeValidator)

fillTxInputWitnesses :: Tx -> TxInput -> TxIn

Translate TxInput to old Plutus.V1.Ledger.Api TxIn taking script and datum witnesses from Tx.

pubKeyTxInput :: TxOutRef -> TxInput

addMintingPolicy :: Versioned MintingPolicy -> (Redeemer, Maybe (Versioned TxOutRef)) -> Tx -> Tx

Add minting policy together with the redeemer into txMintingWitnesses and txScripts accordingly. Doesn't alter txMint.

addScriptTxInput :: TxOutRef -> Versioned Validator -> Redeemer -> Maybe Datum -> Tx -> Tx

Add validator together with the redeemer and datum into txInputs, txData and txScripts accordingly. Datum is optional if the input refers to a script output which contains an inline datum

addReferenceTxInput :: TxOutRef -> Versioned TxOutRef -> Redeemer -> Maybe Datum -> Tx -> Tx

Add script reference together with the redeemer and datum into txInputs and txData accordingly. Datum is optional if the input refers to a script output which contains an inline datum

txRedeemers :: Tx -> Map ScriptPurpose Redeemer

txSpendingRedeemers :: Tx -> Map TxOutRef Redeemer

txMintingRedeemers :: Tx -> Map MintingPolicyHash Redeemer

txRewardingRedeemers :: Tx -> Map Credential Redeemer

txCertifyingRedeemers :: Tx -> Map DCert Redeemer

data Language

Non-Native Script language. This is an Enumerated type. This is expected to be an open type. We will add new Constuctors to this type as additional Non-Native scripting language as are added. We use an enumerated type for two reasons. 1) We can write total functions by case analysis over the constructors 2) We will use DataKinds to make some datatypes indexed by Language For now, the only Non-Native Scriting language is Plutus We might add new languages in the futures.

Note that the the serialization of Language depends on the ordering.

Constructors

PlutusV1

PlutusV2

newtype TxOut

Validator, redeemer, and data scripts of a transaction input that spends a "pay to script" output. inScripts :: Tx -> TxInput -> Maybe (LedgerPlutusVersion, Validator, Redeemer, Datum) inScripts tx i@TxInput{txInputType=TxConsumeScriptAddress pv _ _ _} = case txInType $ fillTxInputWitnesses tx i of Just (ConsumeScriptAddress v r d) -> Just (pv, v, r, d) _ -> Nothing inScripts _ _ = Nothing

Constructors

TxOut

Fields

getTxOut :: TxOut CtxTx BabbageEra

data TxOutRef

A reference to a transaction output. This is a pair of a transaction reference, and an index indicating which of the outputs of that transaction we are referring to.

Constructors

TxOutRef

Fields

txOutRefId :: TxId

txOutRefIdx :: Integer

Index into the referenced transaction's outputs

data Versioned script

A script of some kind with its Plutus language version

Constructors

Versioned

Fields

unversioned :: script

version :: Language