Ledger.Tx.Internal

Documentation

type ReferenceScript = ReferenceScript BabbageEra

type MintingWitnessesMap = Map MintingPolicyHash (Redeemer, Maybe (Versioned TxOutRef))

type ScriptsMap = Map ScriptHash (Versioned Script)

newtype TxOut

Constructors

TxOut

Fields

getTxOut :: TxOut CtxTx BabbageEra

data Certificate

Constructors

Certificate

Fields

certificateDcert :: DCert

certificateRedeemer :: Maybe Redeemer

redeemer for script credential

data Withdrawal

Stake withdrawal, if applicable the script should be included in txScripts.

Constructors

Withdrawal

Fields

withdrawalCredential :: Credential

staking credential

withdrawalAmount :: Integer

amount of withdrawal in Lovelace, must withdraw all eligible amount

withdrawalRedeemer :: Maybe Redeemer

redeemer for script credential

cardanoTxOutValue :: TxOut ctx era -> Value

txOutValue :: TxOut -> Value

outValue :: Lens TxOut TxOut Value (TxOutValue BabbageEra)

outValue' :: Lens' TxOut (TxOutValue BabbageEra)

toSizedTxOut :: TxOut -> Sized (TxOut StandardBabbage)

toCtxUTxOTxOut :: TxOut -> TxOut CtxUTxO BabbageEra

txOutDatumHash :: TxOut -> Maybe DatumHash

Get a hash from the stored TxOutDatum (either directly or by hashing the inlined datum)

txOutDatum :: forall d. FromData d => TxOut -> Maybe d

cardanoTxOutDatumHash :: TxOutDatum CtxUTxO BabbageEra -> Maybe (Hash ScriptData)

txOutPubKey :: TxOut -> Maybe PubKeyHash

txOutAddress :: TxOut -> CardanoAddress

outAddress :: Lens' TxOut (AddressInEra BabbageEra)

outDatumHash :: Lens TxOut TxOut (Maybe DatumHash) (TxOutDatum CtxTx BabbageEra)

txOutReferenceScript :: TxOut -> ReferenceScript

outReferenceScript :: Lens' TxOut ReferenceScript

lookupScript :: ScriptsMap -> ScriptHash -> Maybe (Versioned Script)

lookupValidator :: ScriptsMap -> ValidatorHash -> Maybe (Versioned Validator)

lookupMintingPolicy :: ScriptsMap -> MintingPolicyHash -> Maybe (Versioned MintingPolicy)

lookupStakeValidator :: ScriptsMap -> StakeValidatorHash -> Maybe (Versioned StakeValidator)

emptyTxBodyContent :: TxBodyContent BuildTx BabbageEra

data Language

Constructors

PlutusV1

PlutusV2

newtype TxOut

Constructors

TxOut

Fields

getTxOut :: TxOut CtxTx BabbageEra

data TxOutRef

Constructors

TxOutRef

Fields

txOutRefId :: TxId

txOutRefIdx :: Integer

data Versioned script

Constructors

Versioned

Fields

unversioned :: script

version :: Language