Prettyprinter.Extras

Description

Copied from plutus-ledger-api because not exported

Documentation

newtype PrettyShow a

Newtype wrapper for deriving Pretty via a Show instance

Constructors

PrettyShow

Fields

unPrettyShow :: a

class Pretty a where

Overloaded conversion to Doc.

Laws:

output should be pretty. :-)

Minimal complete definition

pretty

Methods

pretty :: a -> Doc ann

>>> pretty 1 <+> pretty "hello" <+> pretty 1.234
1 hello 1.234

prettyList :: [a] -> Doc ann

prettyList is only used to define the instance Pretty a => Pretty [a]. In normal circumstances only the pretty function is used.

>>> prettyList [1, 23, 456]
[1, 23, 456]

newtype PrettyFoldable f a

Newtype wrapper for deriving Pretty for a Foldable container by calling toList.

Constructors

PrettyFoldable

Fields

unPrettyFoldable :: f a

newtype Tagged (s :: k) b

A Tagged s b value is a value b with an attached phantom type s. This can be used in place of the more traditional but less safe idiom of passing in an undefined value with the type, because unlike an (s -> b), a Tagged s b can't try to use the argument s as a real value.

Moreover, you don't have to rely on the compiler to inline away the extra argument, because the newtype is "free"

Tagged has kind k -> * -> * if the compiler supports PolyKinds, therefore there is an extra k showing in the instance haddocks that may cause confusion.

Constructors

Tagged b

Orphan instances

(KnownSymbol a, Pretty b) => Pretty (Tagged a b)

Instance details

Methods

pretty :: Tagged a b -> Doc ann

prettyList :: [Tagged a b] -> Doc ann